### 2025-09-30 #9113 @editorialbot
[PRE REVIEW]: AutoEmulate v1.0: A PyTorch tool for end-to-end emulation workflows
**Submitting author:** @sgreenbury (<a href="http://orcid.org/0000-0003-4452-2006">Sam F. Greenbury</a>)
**Repository:** https://github.com/alan-turing-institute/autoemulate
**Branch with paper.md** (empty if default branch): joss-v1.0.0
**Version:** v1.1.1
