### 2025-09-30 7ea0406
ghrs: report 09-30-2304-6DE1 for alan-turing-institute/autoemulate
