### 2025-10-01 3995a58
Update autoemulate/core/compare.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
