### 2025-10-01 be8ec90
Update autoemulate/core/compare.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
