### 2025-10-01 9bfd7c7
only test DEFAULT_EMULATORS in test_transformed
