+++ tests/emulators/test_transformed.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 from autoemulate.core.types import DistributionLike, GaussianLike, TensorLike
 from autoemulate.data.utils import set_random_seed
-from autoemulate.emulators import ALL_EMULATORS
+from autoemulate.emulators import DEFAULT_EMULATORS
 from autoemulate.emulators.base import ProbabilisticEmulator
 from autoemulate.emulators.gaussian_process.exact import GaussianProcess
 from autoemulate.emulators.transformed.base import TransformedEmulator
@@ -131,7 +131,7 @@ def run_test(
     [
         get_pytest_param_yof(model, x_t, y_t, o, f)
         for model, x_t, y_t, o, f in itertools.product(
-            ALL_EMULATORS,
+            DEFAULT_EMULATORS,
             [
                 None,
                 [StandardizeTransform(), PCATransform(n_components=3)],
@@ -174,7 +174,7 @@ def test_transformed_emulator(
     [
         get_pytest_param_yo(model, x_t, y_t, o)
         for model, x_t, y_t, o in itertools.product(
-            [emulator for emulator in ALL_EMULATORS if emulator.is_multioutput()],
+            [emulator for emulator in DEFAULT_EMULATORS if emulator.is_multioutput()],
             [
                 None,
                 [StandardizeTransform()],
@@ -222,7 +222,7 @@ def test_transformed_emulator_100_targets(
     [
         get_pytest_param_yo(model, x_t, y_t, o)
         for model, x_t, y_t, o in itertools.product(
-            [emulator for emulator in ALL_EMULATORS if emulator.is_multioutput()],
+            [emulator for emulator in DEFAULT_EMULATORS if emulator.is_multioutput()],
             [
                 None,
                 [StandardizeTransform()],
