### 2025-10-01 21628c2
simplify specification of ALL_EMULATORS
