### 2025-10-01 41df02d
Add MLP with beta distribution output
