### 2025-10-01 8f75271
Revise to ZeroOneInflatedBeta
