+++ autoemulate/emulators/nn/mlp.py
@@ -1,7 +1,6 @@
 import torch
 import torch.nn.functional as F
 from torch import nn
-from torch.distributions import Beta
 
 from autoemulate.core.device import TorchDeviceMixin
 from autoemulate.core.types import DeviceLike, TensorLike
@@ -9,6 +8,105 @@ from autoemulate.data.utils import set_random_seed
 from autoemulate.transforms.standardize import StandardizeTransform
 
 from ..base import DropoutTorchBackend
+
+
+class ZeroOneInflatedBeta(torch.distributions.Distribution):
+    """ZeroOneInflatedBeta."""
+
+    arg_constraints = {  # type: ignore  # noqa: PGH003, RUF012
+        "pi0": torch.distributions.constraints.unit_interval,  # type: ignore  # noqa: PGH003
+        "pi1": torch.distributions.constraints.unit_interval,  # type: ignore  # noqa: PGH003
+        "concentration1": torch.distributions.constraints.positive,  # type: ignore  # noqa: PGH003
+        "concentration0": torch.distributions.constraints.positive,  # type: ignore  # noqa: PGH003
+    }
+    support = torch.distributions.constraints.unit_interval  # type: ignore  # noqa: PGH003
+
+    def __init__(self, pi0, pi1, concentration1, concentration0, validate_args=None):
+        self.pi0 = pi0
+        self.pi1 = pi1
+        self.concentration1 = concentration1
+        self.concentration0 = concentration0
+        self.beta = torch.distributions.Beta(concentration1, concentration0)
+
+        # Ensure pi0 + pi1 < 1 elementwise
+        if ((self.pi0 + self.pi1) > torch.ones_like(self.pi0)).any():
+            msg = "pi0 + pi1 must be <= 1"
+            raise ValueError(msg)
+
+        super().__init__(validate_args=validate_args)
+
+    def log_prob(self, value):
+        """Log prob."""
+        EPS = 1e-12
+        # Ensure value can broadcast with parameters. If value is 1D [N], make it [N, 1]
+        # to align with [N, num_tasks].
+        value_in = value
+        squeeze_back = False
+        while value_in.dim() < self.pi0.dim():
+            value_in = value_in.unsqueeze(-1)
+            squeeze_back = True
+
+        # Clamp continuous values away from boundaries for Beta support
+        v = value_in.clamp(EPS, 1 - EPS)
+
+        # Broadcast pi0, pi1 to value shape
+        pi0_b = self.pi0.expand_as(value_in)
+        pi1_b = self.pi1.expand_as(value_in)
+
+        # Mixture log probs
+        logp0 = torch.log(pi0_b + EPS)
+        logp1 = torch.log(pi1_b + EPS)
+        mix_log = torch.log(1 - pi0_b - pi1_b + EPS)
+        beta_lp = self.beta.log_prob(v)
+
+        cont = mix_log + beta_lp
+        logp = torch.where(
+            value_in == 0, logp0, torch.where(value_in == 1, logp1, cont)
+        )
+
+        if squeeze_back and logp.shape[-1] == 1:
+            logp = logp.squeeze(-1)
+        return logp
+
+    @property
+    def mean(self):
+        """Mixture mean: pi1*1 + (1-pi0-pi1)*beta.mean."""
+        p0 = self.pi0
+        p1 = self.pi1
+        p_cont = 1 - p0 - p1
+        return p1 + p_cont * self.beta.mean
+
+    @property
+    def variance(self):
+        """Mixture variance computed from mixture second moment."""
+        p0 = self.pi0
+        p1 = self.pi1
+        p_cont = 1 - p0 - p1
+        # E[X] for mixture
+        mean = p1 + p_cont * self.beta.mean
[…]
