### 2025-10-01 b978f21
ghrs: report 10-01-2304-6FF0 for alan-turing-institute/autoemulate
