+++ alan-turing-institute/autoemulate/latest-report/report.md
@@ -1,5 +1,5 @@
 % Statistics for alan-turing-institute/autoemulate
-% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-09-30 23:04 UTC.
+% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-10-01 23:04 UTC.
 
 
 ## Views
@@ -7,12 +7,12 @@
 #### Unique visitors
 <div id="chart_views_unique" class="full-width-chart"></div>
 
-Cumulative: 389
+Cumulative: 397
 
 #### Total views
 <div id="chart_views_total" class="full-width-chart"></div>
 
-Cumulative: 4882
+Cumulative: 5023
 
 <div class="pagebreak-for-print"> </div>
 
@@ -21,12 +21,12 @@ Cumulative: 4882
 #### Unique cloners
 <div id="chart_clones_unique" class="full-width-chart"></div>
 
-Cumulative: 794
+Cumulative: 818
 
 #### Total clones
 <div id="chart_clones_total" class="full-width-chart"></div>
 
-Cumulative: 4870
+Cumulative: 5048
 
 
 
@@ -90,16 +90,16 @@ Top 15 referrers: 01: `github.com`, 02: `alan-turing-institute.github.io`, 03: `
 
 <div id="chart_paths_top_n_alltime" class="full-width-chart"></div>
 
-Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/778`, 04: `/pull/776`, 05: `/pull/775`, 06: `/pull/772`, 07: `/pull/775/files`, 08: `/actions/runs/17269709252/job/49077323330`, 09: `/issues`, 10: `/pull/632`
+Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/776`, 04: `/pull/778`, 05: `/pull/775`, 06: `/pull/772`, 07: `/pull/632`, 08: `/issues`, 09: `/pull/775/files`, 10: `/actions/runs/17269709252/job/49077323330`
 
 
 <script type="text/javascript">
[…]
