### 2025-10-02 f2ef5b3
Add supports_uq, fix docstring
