### 2025-10-02 e525b7f
rename kwargs to scheduler_kwargs
