### 2025-10-02 99b4646
Revise API to support passing `parameters_range`
