### 2025-10-02 f63de0f
ghrs: report 10-02-2304-AF5E for alan-turing-institute/autoemulate
