### 2025-10-03 307da0e
update GP constructor function
