+++ autoemulate/emulators/gaussian_process/exact.py
@@ -70,7 +70,7 @@ class GaussianProcess(GaussianProcessEmulator, gpytorch.models.ExactGP):
         lr: float = 2e-1,
         early_stopping: EarlyStopping | None = None,
         device: DeviceLike | None = None,
-        **kwargs,
+        **scheduler_kwargs,
     ):
         """
         Initialize the GaussianProcess emulator.
@@ -108,6 +108,8 @@ class GaussianProcess(GaussianProcessEmulator, gpytorch.models.ExactGP):
         device: DeviceLike | None
             Device to run the model on. If None, uses the default device (usually CPU or
             GPU). Defaults to None.
+        scheduler_kwargs: dict
+            Additional keyword arguments for the learning rate scheduler.
         """
         # Init device
         TorchDeviceMixin.__init__(self, device=device)
@@ -147,7 +149,7 @@ class GaussianProcess(GaussianProcessEmulator, gpytorch.models.ExactGP):
         self.epochs = epochs
         self.lr = lr
         self.optimizer = self.optimizer_cls(self.parameters(), lr=self.lr)  # type: ignore[call-arg] since all optimizers include lr
-        self.scheduler_setup(kwargs)
+        self.scheduler_setup(scheduler_kwargs)
         self.early_stopping = early_stopping
         self.posterior_predictive = posterior_predictive
         self.num_tasks = num_tasks
@@ -462,7 +464,11 @@ GP_REGISTRY = {
 
 
 def create_gp_subclass(
-    name: str, gp_base_class: type[GaussianProcess], **fixed_kwargs
+    name: str,
+    gp_base_class: type[GaussianProcess],
+    covar_module_fn: CovarModuleFn,
+    mean_module_fn: MeanModuleFn,
+    **fixed_kwargs,
 ) -> type[GaussianProcess]:
     """
     Create a subclass of GaussianProcess with given fixed_kwargs.
@@ -476,6 +482,10 @@ def create_gp_subclass(
         Name for the created subclass.
     gp_base_class : type[GaussianProcess]
         Base class to inherit from (typically GaussianProcess).
+    covar_module_fn : CovarModuleFn
+        Covariance module function to use in the subclass.
+    mean_module_fn : MeanModuleFn
+        Mean module function to use in the subclass.
     **fixed_kwargs
         Keyword arguments to fix in the subclass. These parameters will be
         set to the provided values and excluded from hyperparameter tuning.
@@ -497,15 +507,54 @@ def create_gp_subclass(
             f"Use a unique name or delete the existing class from GP_REGISTRY."
         )
 
+    standardize_x = fixed_kwargs.get("standardize_x", False)
+    standardize_y = fixed_kwargs.get("standardize_y", True)
+    fixed_mean_params = fixed_kwargs.get("fixed_mean_params", False)
+    fixed_covar_params = fixed_kwargs.get("fixed_covar_params", False)
+    posterior_predictive = fixed_kwargs.get("posterior_predictive", False)
+    epochs = fixed_kwargs.get("epochs", 50)
+    lr = fixed_kwargs.get("lr", 2e-1)
+    early_stopping = fixed_kwargs.get("early_stopping")
+    device = fixed_kwargs.get("device")
+
     class GaussianProcessSubclass(gp_base_class):
         def __init__(
             self,
-            *args,
-            **kwargs,
+            x: TensorLike,
+            y: TensorLike,
+            standardize_x: bool = standardize_x,
+            standardize_y: bool = standardize_y,
+            likelihood_cls: type[
+                MultitaskGaussianLikelihood
+            ] = MultitaskGaussianLikelihood,
+            mean_module_fn: MeanModuleFn = mean_module_fn,
[…]
