### 2025-10-03 e5d3f0f
add default mean to GP subclass constructor
