### 2025-10-03 7551ce5
rm mean and covar functions from tune params
