+++ alan-turing-institute/autoemulate/latest-report/report.md
@@ -1,5 +1,5 @@
 % Statistics for alan-turing-institute/autoemulate
-% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-10-02 23:04 UTC.
+% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-10-03 23:04 UTC.
 
 
 ## Views
@@ -7,12 +7,12 @@
 #### Unique visitors
 <div id="chart_views_unique" class="full-width-chart"></div>
 
-Cumulative: 414
+Cumulative: 426
 
 #### Total views
 <div id="chart_views_total" class="full-width-chart"></div>
 
-Cumulative: 5141
+Cumulative: 5235
 
 <div class="pagebreak-for-print"> </div>
 
@@ -21,12 +21,12 @@ Cumulative: 5141
 #### Unique cloners
 <div id="chart_clones_unique" class="full-width-chart"></div>
 
-Cumulative: 837
+Cumulative: 856
 
 #### Total clones
 <div id="chart_clones_total" class="full-width-chart"></div>
 
-Cumulative: 5121
+Cumulative: 5275
 
 
 
@@ -90,16 +90,16 @@ Top 15 referrers: 01: `github.com`, 02: `alan-turing-institute.github.io`, 03: `
 
 <div id="chart_paths_top_n_alltime" class="full-width-chart"></div>
 
-Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/778`, 04: `/pull/776`, 05: `/pull/775`, 06: `/pull/772`, 07: `/pull/632`, 08: `/issues`, 09: `/actions/runs/17269709252/job/49077323330`, 10: `/pull/775/files`
+Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/772`, 04: `/pull/776`, 05: `/pull/778`, 06: `/pull/775`, 07: `/pull/632`, 08: `/actions/runs/17269709252/job/49077323330`, 09: `/issues`, 10: `/pull/775/files`
 
 
 <script type="text/javascript">
[…]
