### 2025-10-04 2eb2f6f
ghrs: report 10-04-2304-4C3A for alan-turing-institute/autoemulate
