+++ alan-turing-institute/autoemulate/latest-report/report.md
@@ -1,5 +1,5 @@
 % Statistics for alan-turing-institute/autoemulate
-% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-10-03 23:04 UTC.
+% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-10-04 23:04 UTC.
 
 
 ## Views
@@ -7,12 +7,12 @@
 #### Unique visitors
 <div id="chart_views_unique" class="full-width-chart"></div>
 
-Cumulative: 426
+Cumulative: 431
 
 #### Total views
 <div id="chart_views_total" class="full-width-chart"></div>
 
-Cumulative: 5235
+Cumulative: 5245
 
 <div class="pagebreak-for-print"> </div>
 
@@ -21,12 +21,12 @@ Cumulative: 5235
 #### Unique cloners
 <div id="chart_clones_unique" class="full-width-chart"></div>
 
-Cumulative: 856
+Cumulative: 875
 
 #### Total clones
 <div id="chart_clones_total" class="full-width-chart"></div>
 
-Cumulative: 5275
+Cumulative: 5300
 
 
 
@@ -79,7 +79,7 @@ respond slowly to change (narrow peaks are smoothed out).
 
 <div id="chart_referrers_top_n_alltime" class="full-width-chart"></div>
 
-Top 15 referrers: 01: `github.com`, 02: `alan-turing-institute.github.io`, 03: `autoemulate.com`, 04: `Google`
+Top 15 referrers: 01: `github.com`, 02: `autoemulate.com`, 03: `alan-turing-institute.github.io`, 04: `Google`
 
 
 
@@ -90,16 +90,16 @@ Top 15 referrers: 01: `github.com`, 02: `alan-turing-institute.github.io`, 03: `
 
 <div id="chart_paths_top_n_alltime" class="full-width-chart"></div>
 
-Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/772`, 04: `/pull/776`, 05: `/pull/778`, 06: `/pull/775`, 07: `/pull/632`, 08: `/actions/runs/17269709252/job/49077323330`, 09: `/issues`, 10: `/pull/775/files`
+Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/775`, 04: `/pull/772`, 05: `/pull/776`, 06: `/pull/778`, 07: `/pull/632`, 08: `/issues`, 09: `/pull/775/files`, 10: `/actions/runs/17269709252/job/49077323330`
 
 
 <script type="text/javascript">
[…]
