### 2025-10-05 c4cf511
ghrs: report 10-05-2303-38A9 for alan-turing-institute/autoemulate
