### 2025-10-06 3ce9677
save mlp kwargs in ensembles
