### 2025-10-06 cb44c83
save all MLP and GP input params
