### 2025-10-06 5346644
update HMW to work with Emulator as well Result object, rename result input arg to emulator
