### 2025-10-06 23596b1
add option to pass transformed_emulator_params
