### 2025-10-06 d9503f1
make mlp_kwargs a keyword argument in MLP ensembles
