+++ autoemulate/emulators/gaussian_process/exact.py
@@ -70,7 +70,7 @@ class GaussianProcess(GaussianProcessEmulator, gpytorch.models.ExactGP):
         lr: float = 2e-1,
         early_stopping: EarlyStopping | None = None,
         device: DeviceLike | None = None,
-        **scheduler_kwargs,
+        scheduler_kwargs: dict | None = None,
     ):
         """
         Initialize the GaussianProcess emulator.
@@ -108,7 +108,7 @@ class GaussianProcess(GaussianProcessEmulator, gpytorch.models.ExactGP):
         device: DeviceLike | None
             Device to run the model on. If None, uses the default device (usually CPU or
             GPU). Defaults to None.
-        scheduler_kwargs: dict
+        scheduler_kwargs: dict | None
             Additional keyword arguments for the learning rate scheduler.
         """
         # Init device
@@ -152,7 +152,7 @@ class GaussianProcess(GaussianProcessEmulator, gpytorch.models.ExactGP):
         self.epochs = epochs
         self.lr = lr
         self.optimizer = self.optimizer_cls(self.parameters(), lr=self.lr)  # type: ignore[call-arg] since all optimizers include lr
-        self.scheduler_kwargs = scheduler_kwargs
+        self.scheduler_kwargs = scheduler_kwargs or {}
         self.scheduler_setup(self.scheduler_kwargs)
         self.early_stopping = early_stopping
         self.posterior_predictive = posterior_predictive
