### 2025-10-06 c08f610
add scheduler_cls input keyword arg
