+++ autoemulate/emulators/nn/mlp.py
@@ -1,4 +1,5 @@
 from torch import nn
+from torch.optim.lr_scheduler import LRScheduler
 
 from autoemulate.core.device import TorchDeviceMixin
 from autoemulate.core.types import DeviceLike, TensorLike
@@ -34,6 +35,7 @@ class MLP(DropoutTorchBackend):
         lr: float = 1e-2,
         random_seed: int | None = None,
         device: DeviceLike | None = None,
+        scheduler_cls: type[LRScheduler] | None = None,
         scheduler_kwargs: dict | None = None,
     ):
         """
@@ -77,6 +79,9 @@ class MLP(DropoutTorchBackend):
             Random seed for reproducibility. If None, no seed is set. Defaults to None.
         device: DeviceLike | None
             Device to run the model on (e.g., "cpu", "cuda", "mps"). Defaults to None.
+        scheduler_cls: type[LRScheduler] | None
+            Learning rate scheduler class. If None, no scheduler is used. Defaults to
+            None.
         scheduler_kwargs: dict | None
             Additional keyword arguments related to the scheduler.
 
@@ -126,6 +131,7 @@ class MLP(DropoutTorchBackend):
         self.lr = lr
         self.batch_size = batch_size
         self.optimizer = self.optimizer_cls(self.nn.parameters(), lr=self.lr)  # type: ignore[call-arg] since all optimizers include lr
+        self.scheduler_cls = scheduler_cls
         self.scheduler_kwargs = scheduler_kwargs or {}
         self.scheduler_setup(self.scheduler_kwargs)
         self.to(self.device)
