### 2025-10-06 e0ea75b
check x/y standardization from emulator object
