### 2025-10-06 ae51027
ghrs: report 10-06-2304-5DA2 for alan-turing-institute/autoemulate
