### 2025-10-07 f79b59b
Add limit to python 3.11 for PRs
