+++ .github/workflows/docs.yaml
@@ -11,6 +11,7 @@ permissions:
     contents: write
 jobs:
     docs:
+        if: github.event.pull_request.draft == false || github.event_name == 'push' || github.event_name == 'workflow_dispatch'
         runs-on: ubuntu-latest
         steps:
             - uses: actions/checkout@v4
