### 2025-10-07 0e42855
Add fail fast for CI workflow
