### 2025-10-07 aec14c4
Limit pre-commit workflow to non-draft PRs if PR
