+++ .github/workflows/precommit.yaml
@@ -19,6 +19,7 @@ permissions:
 
 jobs:
     pre-commit:
+        if: github.event_name != 'pull_request' || github.event.pull_request.draft == false
         name: Pre-commit
         runs-on: ubuntu-latest
         steps:
