### 2025-10-07 0bdb67f
Add cancel-in-progress for CI and docs
