+++ .github/workflows/docs.yaml
@@ -7,6 +7,11 @@ on:
     branches:
       - main
   workflow_dispatch:
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 permissions:
     contents: write
 jobs:
