+++ .github/workflows/precommit.yaml
@@ -2,6 +2,7 @@ name: precommit
 on:
     workflow_dispatch:
     pull_request:
+        types: [opened, synchronize, reopened, ready_for_review]
     push:
         branches:
             - main
