### 2025-10-07 199c390
Add maxfail=3 arg for pytest
