### 2025-10-07 5f39acc
Update CI to MSV Python 3.10 for PRs
