### 2025-10-07 25e20b6
Add scripts and config for training
