+++ autoemulate/experimental/run_the_well_experiment.py
@@ -1,0 +1,704 @@
+"""Script to run spatiotemporal emulation experiments from YAML config.
+
+This script provides a command-line interface for running The Well-based
+spatiotemporal emulation experiments using YAML configuration files.
+
+Example usage:
+    # Create an example config
+    python run_the_well_experiment.py --create-example
+
+    # Run an experiment from config
+    python run_the_well_experiment.py --config config.yaml
+
+    # Override output directory
+    python run_the_well_experiment.py --config config.yaml --output-dir ./results
+"""
+
+import argparse
+import logging
+import sys
+from datetime import datetime
+from pathlib import Path
+
+import h5py
+import torch
+from the_well.benchmark.metrics import VRMSE
+from the_well.data import WellDataModule, WellDataset
+from the_well.data.data_formatter import DefaultChannelsFirstFormatter
+
+# Add project root to path if needed
+project_root = Path(__file__).parent.parent.parent
+if str(project_root) not in sys.path:
+    sys.path.insert(0, str(project_root))
+
+from autoemulate.experimental.config_models import (
+    DatasetType,
+    DataSourceType,
+    EmulatorType,
+    ExperimentConfig,
+    FormatterType,
+    LossFunctionType,
+    LRSchedulerType,
+    OptimizerType,
+)
+from autoemulate.experimental.data.spatiotemporal_dataset import (
+    AdvectionDiffusionDataset,
+    AutoEmulateDataModule,
+    BOUTDataset,
+    ReactionDiffusionDataset,
+)
+from autoemulate.experimental.emulators.the_well import (
+    DefaultChannelsFirstFormatterWithTime,
+    TheWellAFNO,
+    TheWellFNO,
+    TheWellFNOWithLearnableWeights,
+    TheWellFNOWithTime,
+    TheWellUNetClassic,
+    TheWellUNetConvNext,
+    TrainerParams,
+)
+from autoemulate.simulations.advection_diffusion import AdvectionDiffusion
+from autoemulate.simulations.reaction_diffusion import ReactionDiffusion
+
+# Set up logger
+logger = logging.getLogger(__name__)
+
+
+def setup_logging(output_dir: Path, log_level: str = "INFO", verbose: bool = False):
+    """Set up logging configuration.
+
+    Parameters
+    ----------
+    output_dir : Path
+        Directory to save log file
+    log_level : str
+        Logging level (DEBUG, INFO, WARNING, ERROR, CRITICAL)
+    verbose : bool
+        If True, also log to console with detailed format
+    """
+    # Create logs directory
+    log_dir = output_dir / "logs"
+    log_dir.mkdir(parents=True, exist_ok=True)
+
+    # Create log filename with timestamp
+    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    log_file = log_dir / f"experiment_{timestamp}.log"
+
+    # Configure root logger
+    root_logger = logging.getLogger()
+    root_logger.setLevel(getattr(logging, log_level.upper()))
+
+    # Remove any existing handlers
+    root_logger.handlers = []
+
+    # File handler - always detailed
+    file_formatter = logging.Formatter(
+        "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+        datefmt="%Y-%m-%d %H:%M:%S",
+    )
+    file_handler = logging.FileHandler(log_file)
+    file_handler.setLevel(logging.DEBUG)
+    file_handler.setFormatter(file_formatter)
+    root_logger.addHandler(file_handler)
+
+    # Console handler
+    if verbose:
+        console_formatter = logging.Formatter(
+            "%(asctime)s - %(levelname)s - %(message)s",
+            datefmt="%H:%M:%S",
+        )
+    else:
[…]
