+++ autoemulate/experimental/README.md
@@ -1,0 +1,478 @@
+# Spatiotemporal Emulation Framework
+
+A unified YAML-based framework for training neural emulators on spatiotemporal data using The Well.
+
+## Quick Start
+
+```bash
+# Run with example configs
+python autoemulate/experimental/run_the_well_experiment.py \
+  --config autoemulate/experimental/configs/turbulent_radiative_layer_2d.yaml
+
+# Create a template config
+python autoemulate/experimental/run_the_well_experiment.py --create-example
+```
+
+## Data Sources (Auto-detected)
+
+The script automatically detects which data source to use:
+
+### 1. The Well Native Datasets
+```yaml
+data:
+  well_dataset_name: "turbulent_radiative_layer_2D"  # Triggers The Well native loading
+  data_path: "../data/the_well/datasets"
+  n_steps_input: 4
+  n_steps_output: 1
+  batch_size: 4
+```
+
+**Example:** `configs/turbulent_radiative_layer_2d.yaml`
+
+### 2. File-Based Data
+```yaml
+data:
+  data_path: "./data/bout"  # Path to HDF5/PyTorch files
+  dataset_type: "bout"
+  n_steps_input: 4
+  n_steps_output: 10
+  batch_size: 4
+```
+
+**Example:** `configs/bout.yaml`
+
+### 3. Generated Data (from Simulators)
+```yaml
+simulator:
+  type: "reaction_diffusion"  # or advection_diffusion
+  parameters_range:
+    feed_rate: [0.02, 0.06]
+    kill_rate: [0.045, 0.065]
+  n: 64
+  T: 100.0
+  dt: 1.0
+
+data:
+  n_train_samples: 200
+  n_valid_samples: 20
+  n_test_samples: 20
+  n_steps_input: 4
+  n_steps_output: 10
+  batch_size: 4
+```
+
+**Examples:** `configs/reaction_diffusion_generated.yaml`, `configs/advection_diffusion_generated.yaml`
+
+## Example Configurations
+
+All configs are in `autoemulate/experimental/configs/`:
+
+| Config | Data Source | Description |
+|--------|-------------|-------------|
+| `turbulent_radiative_layer_2d.yaml` | The Well Native | Turbulent flows with radiation |
+| `bout.yaml` | File | BOUT++ plasma simulation data |
+| `advection_diffusion_generated.yaml` | Generated | Transport phenomena (nu, mu parameters) |
+| `reaction_diffusion_generated.yaml` | Generated | Gray-Scott pattern formation (F, k parameters) |
+
+## Configuration Structure
+
+```yaml
+# Basic info
+experiment_name: "my_experiment"
+description: "What this experiment does"
+
+# Emulator
+emulator_type: "the_well_fno"  # FNO, AFNO, UNet variants
+formatter_type: "default_channels_first"
+
+# Model architecture
+model_params:
+  modes1: 16          # Fourier modes (x-direction)
+  modes2: 16          # Fourier modes (y-direction)
+  width: 32           # Hidden channels
+  n_blocks: 4         # Number of FNO blocks
+
+# Training
+trainer:
+  epochs: 100
+  device: "mps"       # "cuda", "mps", or "cpu"
+  optimizer_type: "adam"
+  optimizer_params:
+    lr: 0.001
+  enable_amp: false   # Mixed precision (CUDA only)
+
+# Teacher forcing (optional)
+teacher_forcing:
+  enabled: true
+  schedule:
+    - start_epoch: 0
+      end_epoch: 50
+      weight: 1.0
+    - start_epoch: 50
+      end_epoch: 100
+      weight: 0.5
+
+# Paths
+paths:
+  output_dir: "./outputs/my_experiment"
[…]
