+++ pyproject.toml
@@ -73,6 +73,7 @@ dev = [
     "hydra-core>=1.3.2",
 ]
 spatiotemporal = [
+    "pydantic>=2.11.10",
     "the-well[dev,benchmark] @ git+https://github.com/PolymathicAI/the_well.git@a123419",
 ]
 

[…]
