### 2025-10-07 9e6538e
Update paths and gitignore
