### 2025-10-07 feee37e
Add initial impl with emulator UQ included in SA
