+++ tests/core/test_sensitivity_analysis.py
@@ -3,6 +3,7 @@ import pandas as pd
 import pytest
 import torch
 from autoemulate.core.sensitivity_analysis import SensitivityAnalysis
+from autoemulate.emulators import GaussianProcessRBF as GaussianProcess
 from autoemulate.emulators.random_forest import RandomForest
 from autoemulate.simulations.projectile import Projectile, ProjectileMultioutput
 
@@ -154,3 +155,199 @@ def test_generate_problem(sa_1d):
     assert problem["num_vars"] == 2
     assert problem["names"] == ["x1", "x2"]
     assert problem["bounds"] == [[0, 2], [0, 2]]
+
+
+# test prediction with variance ------------------------------------------------
+
+
+@pytest.mark.filterwarnings("ignore::gpytorch.utils.warnings.GPInputWarning")
+def test_predict_with_variance(xy_1d):
+    """Test that _predict can return variance for UQ-enabled emulators."""
+    x, y = xy_1d
+    gp = GaussianProcess(x, y)
+    gp.fit(x, y)
+
+    problem = {
+        "num_vars": 2,
+        "names": ["c", "v0"],
+        "bounds": [(-5.0, 1.0), (0.0, 1000.0)],
+    }
+    sa = SensitivityAnalysis(gp, problem=problem)
+
+    # Test with return_variance=False (default)
+    y_pred = sa._predict(x.numpy())
+    assert isinstance(y_pred, np.ndarray)
+    assert y_pred.shape == (100, 1)
+
+    # Test with return_variance=True
+    y_pred, y_var = sa._predict(x.numpy(), return_variance=True)
+    assert isinstance(y_pred, np.ndarray)
+    assert isinstance(y_var, np.ndarray)
+    assert y_pred.shape == (100, 1)
+    assert y_var.shape == (100, 1)
+    assert np.all(y_var >= 0)  # Variance should be non-negative
+
+
+def test_predict_with_variance_no_uq(xy_1d):
+    """Test that _predict returns None variance for non-UQ emulators."""
+    x, y = xy_1d
+    rf = RandomForest(x, y)
+    rf.fit(x, y)
+
+    problem = {
+        "num_vars": 2,
+        "names": ["c", "v0"],
+        "bounds": [(-5.0, 1.0), (0.0, 1000.0)],
+    }
+    sa = SensitivityAnalysis(rf, problem=problem)
+
+    # Test with return_variance=True
+    y_pred, y_var = sa._predict(x.numpy(), return_variance=True)
+    assert isinstance(y_pred, np.ndarray)
+    assert y_var is None
+
+
+# test sensitivity analysis with prediction variance ----------------------------
+
+
+@pytest.mark.filterwarnings("ignore::FutureWarning")
+@pytest.mark.parametrize(
+    ("method", "n_samples", "expected_columns"),
+    [
+        (
+            "sobol",
+            256,
+            ["output", "parameter", "index", "value", "confidence"],
+        ),
+        (
+            "morris",
+            10,
+            ["output", "parameter", "mu", "mu_star", "sigma", "mu_star_conf"],
+        ),
+    ],
+)
+def test_with_prediction_variance(xy_1d, method, n_samples, expected_columns):
+    """Test sensitivity analysis with prediction variance incorporation."""
+    x, y = xy_1d
+    gp = GaussianProcess(x, y)
+    gp.fit(x, y)
+
+    problem = {
+        "num_vars": 2,
+        "names": ["c", "v0"],
+        "bounds": [(-5.0, 1.0), (0.0, 1000.0)],
+    }
+    sa = SensitivityAnalysis(gp, problem=problem)
+
+    # Run with prediction variance
+    df = sa.run(
+        method=method,
[…]
