+++ autoemulate/core/sensitivity_analysis.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Literal
 
 import matplotlib.figure
 import matplotlib.pyplot as plt
@@ -11,6 +12,7 @@ from SALib.analyze.sobol import analyze as sobol_analyze
 from SALib.sample.morris import sample as morris_sample
 from SALib.sample.sobol import sample as sobol_sample
 from SALib.util import ResultDict
+from scipy import stats
 
 from autoemulate.core.plotting import display_figure
 from autoemulate.core.types import NumpyLike, TensorLike
@@ -67,6 +69,32 @@ class SensitivityAnalysis(ConversionMixin):
         self.problem = problem
 
     @staticmethod
+    def _validate_method(method: str) -> Literal["sobol", "morris"]:
+        """
+        Validate and normalize sensitivity analysis method name.
+
+        Parameters
+        ----------
+        method: str
+            Method name (case-insensitive).
+
+        Returns
+        -------
+        Literal["sobol", "morris"]
+            Normalized lowercase method name.
+
+        Raises
+        ------
+        ValueError
+            If method is not 'sobol' or 'morris' (case-insensitive).
+        """
+        method_lower = method.lower()
+        if method_lower not in ["sobol", "morris"]:
+            msg = f"Unknown method: '{method}'. Must be 'sobol' or 'morris'."
+            raise ValueError(msg)
+        return method_lower  # type: ignore[return-value]
+
+    @staticmethod
     def _check_problem(problem: dict) -> dict:
         """Check that the problem definition is valid."""
         if not isinstance(problem, dict):
@@ -119,7 +147,8 @@ class SensitivityAnalysis(ConversionMixin):
             ],
         }
 
-    def _sample(self, method: str, N: int) -> NumpyLike:
+    def _sample(self, method: Literal["sobol", "morris"], N: int) -> NumpyLike:
+        method = self._validate_method(method)
         if method == "sobol":
             # Saltelli sampling
             return sobol_sample(self.problem, N)
@@ -129,7 +158,9 @@ class SensitivityAnalysis(ConversionMixin):
         msg = f"Unknown method: {method}. Must be 'sobol' or 'morris'."
         raise ValueError(msg)
 
-    def _predict(self, param_samples: NumpyLike) -> NumpyLike:
+    def _predict(
+        self, param_samples: NumpyLike, return_variance: bool = False
+    ) -> NumpyLike | tuple[NumpyLike, NumpyLike | None]:
         """
         Make predictions with emulator for parameter samples.
 
@@ -137,14 +168,28 @@ class SensitivityAnalysis(ConversionMixin):
         ----------
         param_samples: NumpyLike
             Array of parameter samples.
+        return_variance: bool
+            Whether to also return prediction variance. Defaults to False.
 
         Returns
         -------
-        NumpyLike
-            Array of emulator predictions.
+        NumpyLike | tuple[NumpyLike, NumpyLike | None]
+            If return_variance is False: Array of emulator predictions.
+            If return_variance is True: Tuple of (predictions, variances).
+            Variances are None if emulator doesn't support UQ.
         """
         param_tensor = self._convert_to_tensors(param_samples)
         assert isinstance(param_tensor, TensorLike)
+
+        if return_variance:
+            y_pred, y_var = self.emulator.predict_mean_and_variance(param_tensor)
+            y_pred_np, _ = self._convert_to_numpy(y_pred)
+            if y_var is not None:
+                y_var_np, _ = self._convert_to_numpy(y_var)
+            else:
+                y_var_np = None
+            return y_pred_np, y_var_np
+
         y_pred = self.emulator.predict_mean(param_tensor)
[…]
