### 2025-10-07 418f082
ghrs: report 10-07-2304-0422 for alan-turing-institute/autoemulate
