### 2025-10-08 5009d1d
update the advection-diffusion config
