### 2025-10-08 570ff40
Fix types and lints
