+++ autoemulate/experimental/run_the_well_experiment.py
@@ -17,21 +17,15 @@ Example usage:
 import argparse
 import logging
 import sys
+from collections.abc import Callable
 from datetime import datetime
 from pathlib import Path
+from typing import Any
 
 import h5py
 import torch
-from the_well.benchmark.metrics import VRMSE
-from the_well.data import WellDataModule, WellDataset
-from the_well.data.data_formatter import DefaultChannelsFirstFormatter
-
-# Add project root to path if needed
-project_root = Path(__file__).parent.parent.parent
-if str(project_root) not in sys.path:
-    sys.path.insert(0, str(project_root))
-
 from autoemulate.experimental.config_models import (
+    DataConfig,
     DatasetType,
     DataSourceType,
     EmulatorType,
@@ -39,7 +33,11 @@ from autoemulate.experimental.config_models import (
     FormatterType,
     LossFunctionType,
     LRSchedulerType,
+    ModelParamsConfig,
     OptimizerType,
+    PathsConfig,
+    SimulatorConfig,
+    TrainerConfig,
 )
 from autoemulate.experimental.data.spatiotemporal_dataset import (
     AdvectionDiffusionDataset,
@@ -50,6 +48,7 @@ from autoemulate.experimental.data.spatiotemporal_dataset import (
 from autoemulate.experimental.emulators.the_well import (
     DefaultChannelsFirstFormatterWithTime,
     TheWellAFNO,
+    TheWellEmulator,
     TheWellFNO,
     TheWellFNOWithLearnableWeights,
     TheWellFNOWithTime,
@@ -59,6 +58,15 @@ from autoemulate.experimental.emulators.the_well import (
 )
 from autoemulate.simulations.advection_diffusion import AdvectionDiffusion
 from autoemulate.simulations.reaction_diffusion import ReactionDiffusion
+from the_well.benchmark.metrics import VRMSE
+from the_well.data import WellDataModule, WellDataset
+from the_well.data.data_formatter import DefaultChannelsFirstFormatter
+
+# Add project root to path if needed
+project_root = Path(__file__).parent.parent.parent
+if str(project_root) not in sys.path:
+    sys.path.insert(0, str(project_root))
+
 
 # Set up logger
 logger = logging.getLogger(__name__)
@@ -115,7 +123,8 @@ def setup_logging(output_dir: Path, log_level: str = "INFO", verbose: bool = Fal
     console_handler.setFormatter(console_formatter)
     root_logger.addHandler(console_handler)
 
-    logger.info(f"Logging initialized. Log file: {log_file}")
+    msg = f"Logging initialized. Log file: {log_file}"
+    logger.info(msg)
 
     return log_file
 
@@ -138,7 +147,7 @@ def create_simulator(config: ExperimentConfig):
             T=sim_cfg.T,
             dt=sim_cfg.dt,
         )
-    elif sim_cfg.type.value == "reaction_diffusion":
+    if sim_cfg.type.value == "reaction_diffusion":
         return ReactionDiffusion(
             parameters_range=sim_cfg.parameters_range,
             output_names=sim_cfg.output_names,
@@ -169,7 +178,9 @@ def get_dataset_class(dataset_type: DatasetType):
     return dataset_classes[dataset_type]
 
 
-def generate_or_load_data(config: ExperimentConfig, simulator=None):
+def generate_or_load_data(
+    config: ExperimentConfig, simulator=None
+) -> WellDataModule | AutoEmulateDataModule:
     """Generate or load data and create data module.
 
     Supports three data source types:
@@ -200,8 +211,6 @@ def generate_or_load_data(config: ExperimentConfig, simulator=None):
             data_cfg.n_steps_input,
             data_cfg.n_steps_output,
         )
-
-        from the_well.data import WellDataModule, WellDataset
 
[…]
