### 2025-10-08 a79261f
Revise to use predicted distribution directly
This enables the predicted distribution to be sampled
from directly when constructing bootstrap samples
and can therefore capture covariance if the predicted
distribution supports this.
