+++ autoemulate/core/sensitivity_analysis.py
@@ -5,6 +5,7 @@ import matplotlib.figure
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import torch
 from matplotlib.axes import Axes
 from matplotlib.lines import Line2D
 from SALib.analyze.morris import analyze as morris_analyze
@@ -15,7 +16,7 @@ from SALib.util import ResultDict
 from scipy import stats
 
 from autoemulate.core.plotting import display_figure
-from autoemulate.core.types import NumpyLike, TensorLike
+from autoemulate.core.types import DistributionLike, NumpyLike, TensorLike
 from autoemulate.data.utils import ConversionMixin
 from autoemulate.emulators.base import Emulator
 
@@ -160,7 +161,7 @@ class SensitivityAnalysis(ConversionMixin):
 
     def _predict(
         self, param_samples: NumpyLike, return_variance: bool = False
-    ) -> NumpyLike | tuple[NumpyLike, NumpyLike | None]:
+    ) -> NumpyLike | DistributionLike:
         """
         Make predictions with emulator for parameter samples.
 
@@ -182,13 +183,12 @@ class SensitivityAnalysis(ConversionMixin):
         assert isinstance(param_tensor, TensorLike)
 
         if return_variance:
-            y_pred, y_var = self.emulator.predict_mean_and_variance(param_tensor)
-            y_pred_np, _ = self._convert_to_numpy(y_pred)
-            if y_var is not None:
-                y_var_np, _ = self._convert_to_numpy(y_var)
-            else:
-                y_var_np = None
-            return y_pred_np, y_var_np
+            if not self.emulator.supports_uq:
+                msg = "Emulator does not support uncertainty quantification."
+                raise ValueError(msg)
+            y_pred = self.emulator.predict(param_tensor)
+            assert isinstance(y_pred, DistributionLike), "Expected DistributionLike"
+            return y_pred
 
         y_pred = self.emulator.predict_mean(param_tensor)
         y_pred_np, _ = self._convert_to_numpy(y_pred)
@@ -258,8 +258,7 @@ class SensitivityAnalysis(ConversionMixin):
         self,
         method: Literal["sobol", "morris"],
         param_samples: NumpyLike,
-        y_mean: NumpyLike,
-        y_var: NumpyLike,
+        y_dist: DistributionLike,
         output_names: list[str],
         conf_level: float,
         n_bootstrap: int,
@@ -271,25 +270,20 @@ class SensitivityAnalysis(ConversionMixin):
         bootstrap_results: dict[str, dict[str, list]] = {
             name: {metric: [] for metric in metrics} for name in output_names
         }
-
-        for _ in range(n_bootstrap):
-            y_sampled = np.random.normal(
-                loc=y_mean,
-                scale=np.sqrt(np.maximum(y_var, 1e-10)),
-            )
+        y_samples = y_dist.sample(  # (n_bootstrap, n_points, n_outputs)
+            torch.Size([n_bootstrap])
+        )
+        for m in range(n_bootstrap):
+            y_sample = y_samples[m].numpy()  # (n_points, n_outputs)
 
             for i, name in enumerate(output_names):
+                y = y_sample[:, i]  # (n_points,)
                 if config["needs_param_samples"]:
                     Si = config["analyze_fn"](
-                        self.problem,
-                        param_samples,
-                        y_sampled[:, i],
-                        conf_level=conf_level,
+                        self.problem, param_samples, y, conf_level=conf_level
                     )
                 else:
-                    Si = config["analyze_fn"](
-                        self.problem, y_sampled[:, i], conf_level=conf_level
-                    )
+                    Si = config["analyze_fn"](self.problem, y, conf_level=conf_level)
[…]
