### 2025-10-08 8439e4a
fix scheduler kwarg passing to scheduler_setup
