+++ autoemulate/emulators/gaussian_process/exact.py
@@ -158,7 +158,7 @@ class GaussianProcess(GaussianProcessEmulator, gpytorch.models.ExactGP):
         self.optimizer = self.optimizer_cls(self.parameters(), lr=self.lr)  # type: ignore[call-arg] since all optimizers include lr
         self.scheduler_cls = scheduler_cls
         self.scheduler_kwargs = scheduler_kwargs or {}
-        self.scheduler_setup(self.scheduler_kwargs)
+        self.scheduler_setup(**self.scheduler_kwargs)
         self.early_stopping = early_stopping
         self.posterior_predictive = posterior_predictive
         self.num_tasks = num_tasks
