### 2025-10-08 38283a8
update scheduler_setup method
