+++ autoemulate/emulators/polynomials.py
@@ -87,7 +87,7 @@ class PolynomialRegression(PyTorchBackend):
         self.optimizer = self.optimizer_cls(self.linear.parameters(), lr=self.lr)  # type: ignore[call-arg] since all optimizers include lr
         self.scheduler_cls = scheduler_cls
         self.scheduler_kwargs = scheduler_kwargs or {}
-        self.scheduler_setup(**self.scheduler_kwargs)
+        self.scheduler_setup(self.scheduler_kwargs)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """Forward pass through for polynomial regression."""
