### 2025-10-08 3c523a6
update scheduler tests
