### 2025-10-08 f6a2377
add reinitialize method
