### 2025-10-08 5b0d9f2
add reinitialize method
