### 2025-10-08 4102bbd
add tensor conversion and device handling to TransformedEmulator
