### 2025-10-08 36a8b64
update Emulator.fit to expect InputLike, not TensorLike
