### 2025-10-08 12371a6
refactor fit_from_reinitialised function
