### 2025-10-08 aa18dca
use fit_from_initialized in learners
