### 2025-10-08 dee04f5
revert changes in learners
