### 2025-10-08 e6d12fc
accept both emulator and result as keyword args to HMW
make sure changes dont break API
