### 2025-10-09 1994b2f
WIP adding general torchmetrics support
