### 2025-10-09 a118e8b
disable ruff qa for file
