+++ autoemulate/emulators/nn/mlp.py
@@ -1,13 +1,11 @@
-import torch
-from torch import nn, optim
+from torch import nn
 
 from autoemulate.core.device import TorchDeviceMixin
-from autoemulate.core.types import DeviceLike, GaussianLike, TensorLike
+from autoemulate.core.types import DeviceLike, TensorLike
 from autoemulate.data.utils import set_random_seed
 from autoemulate.transforms.standardize import StandardizeTransform
-from autoemulate.transforms.utils import make_positive_definite
 
-from ..base import DropoutTorchBackend, GaussianEmulator
+from ..base import DropoutTorchBackend
 
 
 class MLP(DropoutTorchBackend):
@@ -146,125 +144,4 @@ class MLP(DropoutTorchBackend):
             "dropout_prob": [0.3, None],
             "scheduler_cls": scheduler_params["scheduler_cls"],
             "scheduler_kwargs": scheduler_params["scheduler_kwargs"],
-        }
-
-
-class GaussianMLP(DropoutTorchBackend, GaussianEmulator):
-    """Multi-Layer Perceptron (MLP) emulator with Gaussian outputs."""
-
-    def __init__(
-        self,
-        x,
-        y,
-        activation_cls=nn.ReLU,
-        loss_fn_cls=nn.MSELoss,
-        optimizer_cls=optim.Adam,
-        scheduler_cls=None,
-        scheduler_kwargs=None,
-        epochs=100,
-        layer_dims=None,
-        weight_init="default",
-        scale=1,
-        bias_init="default",
-        dropout_prob=None,
-        lr=1e-2,
-        random_seed=None,
-        device=None,
-        **kwargs,
-    ):
-        TorchDeviceMixin.__init__(self, device=device)
-        nn.Module.__init__(self)
-
-        if random_seed is not None:
-            set_random_seed(seed=random_seed)
-
-        # Ensure x and y are tensors with correct dimensions
-        x, y = self._convert_to_tensors(x, y)
-
-        # Construct the MLP layers
-        # Total params required for last layer: mean + tril covariance
-        num_params = y.shape[1] + (y.shape[1] * (y.shape[1] + 1)) // 2
-        layer_dims = (
-            [x.shape[1], *layer_dims]
-            if layer_dims
-            else [x.shape[1], 4 * num_params, 2 * num_params]
-        )
-        layers = []
-        for idx, dim in enumerate(layer_dims[1:]):
-            layers.append(nn.Linear(layer_dims[idx], dim, device=self.device))
-            layers.append(activation_cls())
-            if dropout_prob is not None:
-                layers.append(nn.Dropout(p=dropout_prob))
-
-        # Add final layer without activation
-        layers.append(nn.Linear(layer_dims[-1], num_params, device=self.device))
-        self.nn = nn.Sequential(*layers)
-
-        # Finalize initialization
-        self._initialize_weights(weight_init, scale, bias_init)
-        self.epochs = epochs
-        self.loss_fn = loss_fn_cls()
-        self.num_tasks = y.shape[1]
-        self.optimizer = optimizer_cls(self.nn.parameters(), lr=lr)
-        self.scheduler_cls = scheduler_cls
-        self.scheduler_setup(scheduler_kwargs)
-        self.to(device)
-
-    def _predict(self, x, with_grad=False):
-        """Predict using the MLP model."""
-        with torch.set_grad_enabled(with_grad):
-            self.nn.eval()
-            return self(x)
-
-    def forward(self, x):
-        """Forward pass for the Gaussian MLP."""
-        y = self.nn(x)
-        mean = y[..., : self.num_tasks]
-
-        # Use Cholesky decomposition to guarantee PSD covariance matrix
[…]
