### 2025-10-09 a21c82f
Add torchmetrics support for tuning and evaluation
