+++ autoemulate/core/results.py
@@ -1,7 +1,12 @@
+import logging
+
 import pandas as pd
 
+from autoemulate.core.metrics import AVAILABLE_METRICS
 from autoemulate.core.types import ModelParams
 from autoemulate.emulators.transformed.base import TransformedEmulator
+
+logger = logging.getLogger("autoemulate")
 
 
 class Result:
@@ -13,14 +18,8 @@ class Result:
         model_name: str,
         model: TransformedEmulator,
         params: ModelParams,
-        r2_test: float,
-        rmse_test: float,
-        r2_train: float,
-        rmse_train: float,
-        r2_test_std: float,
-        rmse_test_std: float,
-        r2_train_std: float,
-        rmse_train_std: float,
+        test_metrics: dict[str, tuple[float, float]],
+        train_metrics: dict[str, tuple[float, float]],
     ):
         """Initialize a Result object.
 
@@ -34,22 +33,12 @@ class Result:
             The emulator model used for predictions.
         params: ModelParams
             Parameters used for the model.
-        r2_test: float
-            R2 score on the test set.
-        rmse_test: float
-            Root Mean Squared Error on the test set.
-        r2_train: float
-            R2 score on the training set.
-        rmse_train: float
-            Root Mean Squared Error on the training set.
-        r2_test_std: float
-            Standard deviation of the R2 score on the test set.
-        rmse_test_std: float
-            Standard deviation of the RMSE on the test set.
-        r2_train_std: float
-            Standard deviation of the R2 score on the training set.
-        rmse_train_std: float
-            Standard deviation of the RMSE on the training set.
+        test_metrics: dict[str, tuple[float, float]]
+            Dictionary of metrics on the test set. Each key is a metric name and
+            each value is a tuple of (mean, std).
+        train_metrics: dict[str, tuple[float, float]]
+            Dictionary of metrics on the training set. Each key is a metric name and
+            each value is a tuple of (mean, std).
 
         """
         self.id = id
@@ -58,14 +47,8 @@ class Result:
         self.x_transforms = model.x_transforms
         self.y_transforms = model.y_transforms
         self.params = params
-        self.r2_test = r2_test
-        self.rmse_test = rmse_test
-        self.r2_test_std = r2_test_std
-        self.rmse_test_std = rmse_test_std
-        self.r2_train = r2_train
-        self.rmse_train = rmse_train
-        self.r2_train_std = r2_train_std
-        self.rmse_train_std = rmse_train_std
+        self.test_metrics = test_metrics
+        self.train_metrics = train_metrics
 
     def metadata_df(self) -> pd.DataFrame:
         """
@@ -75,9 +58,8 @@ class Result:
         -------
         pd.DataFrame
             DataFrame with columns:
-            ['id', 'model_name', 'x_transforms', 'y_transforms', 'params',
-            'r2_test', 'rmse_test', 'r2_test_std', 'rmse_test_std',
-            'r2_train', 'rmse_train', 'r2_train_std', 'rmse_train_std'].
+            ['id', 'model_name', 'x_transforms', 'y_transforms', 'params']
+            plus columns for each metric in test_metrics and train_metrics.
         """
 
         # Serialize the params dictionary to a string representation
@@ -90,23 +72,25 @@ class Result:
                     out[k] = v
             return out
 
-        return pd.DataFrame(
-            {
-                "id": [self.id],
-                "model_name": [self.model_name],
-                "x_transforms": [self.x_transforms],
-                "y_transforms": [self.y_transforms],
-                "params": str(serialize_params(self.params)),
-                "r2_test": [self.r2_test],
-                "rmse_test": [self.rmse_test],
-                "r2_test_std": [self.r2_test_std],
[…]
