### 2025-10-09 fcd2827
revert Emulator data input types to TensorLike (not InputLike)
