### 2025-10-09 c10826f
rm now unnecessary tensor transform from TransformedEmulator
