### 2025-10-09 41b0972
rm now unnecessary tensor transform from Emulator
