### 2025-10-09 e602188
update fit_from_reinitialize to expect TensorLike not InputLike
