### 2025-10-09 3c8d3dd
ensure tensor is float
