### 2025-10-09 61db6eb
ensure tensors are floats
