### 2025-10-09 0c238b0
revert scheduler_params back to kwargs
