### 2025-10-09 b77e283
use convert_to_tensors method
