### 2025-10-09 06c7ca0
rename scheduler_kwargs to scheduler_params
