+++ autoemulate/emulators/polynomials.py
@@ -31,7 +31,7 @@ class PolynomialRegression(PyTorchBackend):
         random_seed: int | None = None,
         device: DeviceLike | None = None,
         scheduler_cls: type[LRScheduler] | None = None,
-        scheduler_kwargs: dict | None = None,
+        scheduler_params: dict | None = None,
     ):
         """Initialize a PolynomialRegression emulator.
 
@@ -61,7 +61,7 @@ class PolynomialRegression(PyTorchBackend):
         scheduler_cls: type[LRScheduler] | None
             Learning rate scheduler class. If None, no scheduler is used. Defaults to
             None.
-        scheduler_kwargs: dict | None
+        scheduler_params: dict | None
             Additional keyword arguments related to the scheduler.
         """
         super().__init__()
@@ -86,8 +86,8 @@ class PolynomialRegression(PyTorchBackend):
         ).to(self.device)
         self.optimizer = self.optimizer_cls(self.linear.parameters(), lr=self.lr)  # type: ignore[call-arg] since all optimizers include lr
         self.scheduler_cls = scheduler_cls
-        self.scheduler_kwargs = scheduler_kwargs or {}
-        self.scheduler_setup(self.scheduler_kwargs)
+        self.scheduler_params = scheduler_params or {}
+        self.scheduler_setup(self.scheduler_params)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """Forward pass through for polynomial regression."""
@@ -103,11 +103,11 @@ class PolynomialRegression(PyTorchBackend):
     @staticmethod
     def get_tune_params():
         """Return a dictionary of hyperparameters to tune."""
-        scheduler_params = PolynomialRegression.scheduler_params()
+        scheduler_specs = PolynomialRegression.get_scheduler_params()
         return {
             "lr": [1e-3, 1e-2, 1e-1, 2e-1],
             "epochs": [50, 100, 200, 500, 1000],
             "batch_size": [8, 16, 32],
-            "scheduler_cls": scheduler_params["scheduler_cls"],
-            "scheduler_kwargs": scheduler_params["scheduler_kwargs"],
+            "scheduler_cls": scheduler_specs["scheduler_cls"],
+            "scheduler_params": scheduler_specs["scheduler_params"],
         }
