### 2025-10-09 de55b88
ghrs: report 10-09-2304-E0BD for alan-turing-institute/autoemulate
