### 2025-10-10 2fb3f8d
Fix vmap incompatibility
