### 2025-10-11 4c1b735
ghrs: report 10-11-2304-9138 for alan-turing-institute/autoemulate
