+++ alan-turing-institute/autoemulate/latest-report/report.md
@@ -1,5 +1,5 @@
 % Statistics for alan-turing-institute/autoemulate
-% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-10-11 23:04 UTC.
+% Generated for [alan-turing-institute/autoemulate](https://github.com/alan-turing-institute/autoemulate) with [jgehrcke/github-repo-stats](https://github.com/jgehrcke/github-repo-stats) at 2025-10-12 23:04 UTC.
 
 
 ## Views
@@ -21,12 +21,12 @@ Cumulative: 5737
 #### Unique cloners
 <div id="chart_clones_unique" class="full-width-chart"></div>
 
-Cumulative: 1063
+Cumulative: 1074
 
 #### Total clones
 <div id="chart_clones_total" class="full-width-chart"></div>
 
-Cumulative: 5791
+Cumulative: 5806
 
 
 
@@ -90,16 +90,16 @@ Top 15 referrers: 01: `github.com`, 02: `autoemulate.com`, 03: `alan-turing-inst
 
 <div id="chart_paths_top_n_alltime" class="full-width-chart"></div>
 
-Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/776`, 04: `/pull/778`, 05: `/pull/775`, 06: `/pull/772`, 07: `/issues`, 08: `/pull/632`, 09: `/pull/775/files`, 10: `/actions/runs/17269709252/job/49077323330`
+Top 15 paths: 01: `/`, 02: `/pulls`, 03: `/pull/776`, 04: `/pull/772`, 05: `/pull/775`, 06: `/pull/778`, 07: `/issues`, 08: `/pull/632`, 09: `/actions/runs/17269709252/job/49077323330`, 10: `/pull/775/files`
 
 
 <script type="text/javascript">
[…]
