### 2025-10-13 39797ad
use fit_from_initialized in AL, change types to DistributionLike from GaussianLike to match TranformedEmulator predict type
