### 2025-10-14 8498cb3
avoid code repetition
