### 2025-10-14 85bf20d
revert to emulator.fit instead of fit_from_reinitialised in stream based AL
