+++ autoemulate/learners/base.py
@@ -7,7 +7,6 @@ from anytree import Node, RenderTree
 from torcheval.metrics import MeanSquaredError, R2Score
 
 from autoemulate.core.logging_config import get_configured_logger
-from autoemulate.core.reinitialize import fit_from_reinitialized
 from autoemulate.data.utils import ValidationMixin
 from autoemulate.emulators.base import Emulator
 from autoemulate.simulations.base import Simulator
@@ -48,12 +47,7 @@ class Learner(ValidationMixin, ABC):
         log_level = getattr(self, "log_level", "progress_bar")
         self.logger, self.progress_bar = get_configured_logger(log_level)
         self.logger.info("Initializing Learner with training data.")
-        self.emulator = fit_from_reinitialized(
-            self.x_train,
-            self.y_train,
-            emulator=self.emulator,
-            device=self.emulator.device,
-        )
+        self.emulator.fit(self.x_train, self.y_train)
         self.logger.info("Emulator fitted with initial training data.")
         self.in_dim = self.x_train.shape[1]
         self.out_dim = self.y_train.shape[1]
@@ -162,12 +156,7 @@ class Active(Learner):
             assert isinstance(y_true, TensorLike)
             self.x_train = torch.cat([self.x_train, x])
             self.y_train = torch.cat([self.y_train, y_true])
-            self.emulator = fit_from_reinitialized(
-                self.x_train,
-                self.y_train,
-                emulator=self.emulator,
-                device=self.emulator.device,
-            )
+            self.emulator.fit(self.x_train, self.y_train)
             self.mse.update(y_pred, y_true)
             self.r2.update(y_pred, y_true)
             self.n_queries += 1
