### 2025-10-14 704fe37
add option to change whether fit from reinitialized or not in AL
