### 2025-10-14 b5b902d
Update autoemulate/calibration/history_matching.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
