### 2025-10-14 bcd2939
Update autoemulate/calibration/history_matching.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
