### 2025-10-14 bbccf27
ghrs: report 10-14-2304-F10D for alan-turing-institute/autoemulate
