### 2025-10-15 6002c07
Fix docstrings
Co-authored-by: Radka Jersakova <r.jersakova@gmail.com>
