### 2025-10-15 597977d
raise error if output distribution does not have a variance property
