+++ autoemulate/core/compare.py
@@ -406,14 +406,18 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
                                     "parameters",
                                     model_cls.__name__,
                                 )
-                                # extract default parameters from the model's __init__
+                                # Extract default parameters from the model's __init__
                                 init_sig = inspect.signature(model_cls.__init__)
-                                init_params = {
+                                default_params = {
                                     param_name: param.default
                                     for param_name, param in init_sig.parameters.items()
                                     if param_name in model_cls.get_tune_params()
                                 }
-                                best_params_for_this_model = init_params
+                                # Overwrite defaults with user-supplied values
+                                best_params_for_this_model = {
+                                    **default_params,
+                                    **self.model_params,
+                                }
 
                             self.logger.debug(
                                 'Running cross-validation for model "%s" '
