### 2025-10-15 dc4e77c
Remove method as in base class
