### 2025-10-15 d553bc1
Add back required _predict given base class structure
