### 2025-10-15 58d5d8e
Add optional full covariance
