### 2025-10-16 e00541a
update type hinting
