+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "autoemulate"
-version = "1.1.2"
+version = "1.2.0"
 description = "A python package for semi-automated emulation"
 license = "MIT"
 authors = [{ name = "AutoEmulate contributors (see our GitHub page)" }]
