+++ pyproject.toml
@@ -75,6 +75,7 @@ dev = [
 spatiotemporal = [
     "pydantic>=2.11.10",
     "the-well[dev,benchmark] @ git+https://github.com/PolymathicAI/the_well.git@a123419",
+    # "triton>=3.0.0",
 ]
 
 [tool.setuptools]
@@ -168,5 +169,6 @@ convention = "numpy"
 # Platform-specific triton handling:
 # - On macOS arm64: fetch triton from the GitHub repo and build from source (uv/pip will attempt to build)
 override-dependencies = [
-    "triton @ git+https://github.com/openai/triton.git@main; sys_platform == 'darwin' and platform_machine == 'arm64'",
+    #"triton @ git+https://github.com/openai/triton.git@main; sys_platform == 'darwin' and platform_machine == 'arm64'",
+    "triton @ git+https://github.com/openai/triton.git@main",
 ]
