### 2025-10-16 cd8e07c
Update advection-diffusion config
