+++ autoemulate/core/plotting.py
@@ -5,7 +5,7 @@ from IPython.core.getipython import get_ipython
 from matplotlib.axes import Axes
 from matplotlib.figure import Figure
 
-from autoemulate.core.types import NumpyLike, TensorLike
+from autoemulate.core.types import DistributionLike, NumpyLike, TensorLike
 from autoemulate.emulators.base import Emulator
 
 
@@ -236,7 +236,6 @@ def mean_and_var_surface(
         The predicted variance on the grid.
     grid: list[TensorLike]
         The grid of parameter values used for predictions.
-
     """
     # Determine which parameters to vary and which to fix
     grid_params = {}
@@ -412,8 +411,141 @@ def create_and_plot_slice(
         param_pair_names,
         vmin,
         vmax,
-        fixed_params_info=f"{', '.join(fixed_params)} at {quantile:.1f} quantile"
-        if len(fixed_params) > 0
-        else "None",
+        fixed_params_info=(
+            f"{', '.join(fixed_params)} at {quantile:.1f} quantile"
+            if len(fixed_params) > 0
+            else "None"
+        ),
     )
+    return fig, ax
+
+
+def coverage_from_distributions(
+    y_pred: DistributionLike,
+    y_true: TensorLike,
+    levels: list[float] | NumpyLike | TensorLike | None = None,
+    n_samples: int = 2000,
+    joint: bool = False,
+) -> tuple[NumpyLike, NumpyLike]:
+    """Compute empirical coverage for a set of nominal confidence levels.
+
+    Parameters
+    ----------
+    y_pred: DistributionLike
+        The emulator predicted distribution.
+    y_true: TensorLike
+        The true values.
+    levels: array-like, optional
+        Nominal coverage levels (between 0 and 1). If None, a default grid is
+        used. Defaults to None.
+    n_samples: int
+        Number of Monte-Carlo samples to draw from the predictive
+        distribution.
+    joint: bool
+        If True and the predictive outputs are multivariate, compute joint
+        coverage (i.e., the true vector must lie inside the interval for all
+        dimensions). If False (default), compute marginal coverage per output
+        dimension and return the mean across data points.
+
+    Returns
+    -------
+    levels: np.ndarray
+        Nominal coverage levels.
+    empirical: np.ndarray
+        Empirical coverages. Shape is (len(levels), output_dim) when
+        `joint=False` and output_dim>1, or (len(levels),) when joint=True or
+        output_dim==1.
+    """
+    if levels is None:
+        levels = np.linspace(0.01, 0.99, 49)
+    levels = np.asarray(levels)
+
+    samples = y_pred.sample((n_samples,))
+
+    # compute empirical intervals using sample quantiles
+    empirical_list = []
+    for p in levels:
+        lower_q = (1.0 - p) / 2.0
+        upper_q = 1.0 - lower_q
+        lower = torch.quantile(samples, float(lower_q), dim=0)
+        upper = torch.quantile(samples, float(upper_q), dim=0)
+        inside = (y_true >= lower) & (y_true <= upper)
+        if joint:
+            inside_all = inside.all(dim=-1)
+            empirical = inside_all.float().mean().item()
+        else:
+            # marginal per-dim coverage
+            empirical = inside.float().mean(dim=0).cpu().numpy()
+        empirical_list.append(empirical)
+
+    empirical_arr = np.asarray(empirical_list)
+
+    return levels, empirical_arr
+
+
+def plot_calibration_from_distributions(
+    y_pred: DistributionLike,
+    y_true: TensorLike,
+    levels: np.ndarray | None = None,
+    n_samples: int = 2000,
+    joint: bool = False,
+    title: str | None = None,
+    legend: bool = True,
+    figsize: tuple[int, int] | None = None,
+):
[…]
