+++ autoemulate/core/plotting.py
@@ -531,18 +531,18 @@ def plot_calibration_from_distributions(
     fig, ax = plt.subplots(figsize=figsize)
 
     if empirical.shape[1] == 1:
-        ax.plot(empirical, levels, marker="o", label="empirical")
+        ax.plot(levels, empirical, marker="o", label="empirical")
     else:
         # multiple outputs: plot each dimension
         for i in range(empirical.shape[1]):
-            ax.plot(empirical[:, i], levels, marker="o", label=f"dim {i}")
+            ax.plot(levels, empirical[:, i], marker="o", label=f"dim {i}")
 
     # diagonal reference
     ax.plot([0, 1], [0, 1], linestyle="--", color="gray", label="ideal")
     ax.set_xlim(0, 1)
     ax.set_ylim(0, 1)
-    ax.set_xlabel("Predicted probability")
-    ax.set_ylabel("Observed proportion")
+    ax.set_xlabel("Probability interval")
+    ax.set_ylabel("Proportion of data inside interval")
 
     if title:
         ax.set_title(title)
