### 2025-10-16 400bd04
Fix defaults to improve perf
