### 2025-10-16 2848b9c
update docstrings
