### 2025-10-17 24c861c
Update autoemulate/core/plotting.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
