### 2025-10-17 e2b0760
Update autoemulate/core/plotting.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
