### 2025-10-17 d431f1a
update plot labels, add text to tutorial
