### 2025-10-17 d10db95
use icdf to compute intervals when available
