### 2025-10-17 6db8f09
Type hinting errors
