+++ autoemulate/core/types.py
@@ -1,9 +1,11 @@
+from functools import partial
 from typing import Any, TypeAlias
 
 import numpy as np
 import torch
 import torch.utils
 import torch.utils.data
+import torchmetrics
 from gpytorch.distributions import MultitaskMultivariateNormal, MultivariateNormal
 from torch.utils.data import DataLoader
 
@@ -19,6 +21,7 @@ TuneParams: TypeAlias = dict[str, list[ParamLike]]
 ModelParams: TypeAlias = dict[str, ParamLike]
 TransformedEmulatorParams: TypeAlias = dict[str, ParamLike]
 DeviceLike: TypeAlias = str | torch.device
+MetricLike: TypeAlias = type[torchmetrics.Metric] | partial[torchmetrics.Metric]
 
 # Torch dtype's
 TorchScalarDType = (torch.float32, torch.float64, torch.int32, torch.int64)

[…]
