### 2025-10-17 16e5794
Additional tests into the compare.py
