### 2025-10-17 383b8f4
Update config and trainer scripts
