+++ autoemulate/experimental/configs/advection_diffusion_generated.yaml
@@ -38,7 +38,7 @@ data:
   # No data_path - triggers GENERATED source type
   
   # Data generation
-  n_train_samples: 2000
+  n_train_samples: 1000
   n_valid_samples: 20
   n_test_samples: 20
   random_seed: 42
@@ -46,7 +46,7 @@ data:
   # Dataset configuration
   dataset_type: "advection_diffusion"
   n_steps_input: 4        # Number of input timesteps
-  n_steps_output: 10      # Number of output timesteps (rollout length)
+  n_steps_output: 1      # Number of output timesteps (rollout length)
   batch_size: 4           # Batch size for training
   stride: 1               # Stride for sampling
 
