### 2025-10-19 6c77f05
ghrs: report 10-19-2304-C701 for alan-turing-institute/autoemulate
