### 2025-10-20 2dd4098
Add docstring
