### 2025-10-20 d5080b3
Update scheduler handling
