### 2025-10-20 537c46e
Add registry class for emulators
