+++ autoemulate/emulators/__init__.py
@@ -14,46 +14,175 @@ from .random_forest import RandomForest
 from .svm import SupportVectorMachine
 from .transformed.base import TransformedEmulator
 
-DEFAULT_EMULATORS: list[type[Emulator]] = [
-    GaussianProcessMatern32,
-    GaussianProcessRBF,
-    RadialBasisFunctions,
-    PolynomialRegression,
-    MLP,
-    EnsembleMLP,
-]
 
-# listing non pytorch emulators as we do not expect this list to grow
-NON_PYTORCH_EMULATORS: list[type[Emulator]] = [
-    LightGBM,
-    SupportVectorMachine,
-    RandomForest,
-]
+class Registry:
+    """
+    Registry for managing emulators.
 
-ALL_EMULATORS: list[type[Emulator]] = [
-    *DEFAULT_EMULATORS,
-    *NON_PYTORCH_EMULATORS,
-    GaussianProcessCorrelatedMatern32,
-    GaussianProcessCorrelatedRBF,
-    EnsembleMLPDropout,
-]
+    The Registry class maintains collections of emulator classes organized by
+    their properties (e.g., Gaussian Process emulators, PyTorch-based emulators).
+    It provides methods to register new emulators and retrieve them by name.
 
-PYTORCH_EMULATORS: list[type[Emulator]] = [
-    emulator for emulator in ALL_EMULATORS if emulator not in NON_PYTORCH_EMULATORS
-]
-GAUSSIAN_PROCESS_EMULATORS: list[type[Emulator]] = [
-    emulator
-    for emulator in ALL_EMULATORS
-    if issubclass(emulator, GaussianProcessEmulator)
-]
+    Attributes
+    ----------
+    emulators : dict[str, type[Emulator]]
+        Dictionary mapping emulator names to classes.
+    """
 
-EMULATOR_REGISTRY = {em_cls.model_name().lower(): em_cls for em_cls in ALL_EMULATORS}
-EMULATOR_REGISTRY_SHORT_NAME = {em_cls.short_name(): em_cls for em_cls in ALL_EMULATORS}
+    emulators: dict[str, type[Emulator]]
+
+    def __init__(self):
+        self._default_emulators: list[type[Emulator]] = [
+            GaussianProcessMatern32,
+            GaussianProcessRBF,
+            RadialBasisFunctions,
+            PolynomialRegression,
+            MLP,
+            EnsembleMLP,
+        ]
+
+        # listing non pytorch emulators as we do not expect this list to grow
+        self._non_pytorch_emulators: list[type[Emulator]] = [
+            LightGBM,
+            SupportVectorMachine,
+            RandomForest,
+        ]
+
+        self._all_emulators: list[type[Emulator]] = [
+            *self._default_emulators,
+            *self._non_pytorch_emulators,
+            GaussianProcessCorrelatedMatern32,
+            GaussianProcessCorrelatedRBF,
+            EnsembleMLPDropout,
+        ]
+
+        self._pytorch_emulators: list[type[Emulator]] = [
+            emulator
+            for emulator in self._all_emulators
+            if emulator not in self._non_pytorch_emulators
+        ]
+        self._gaussian_process_emulators: list[type[Emulator]] = [
+            emulator
+            for emulator in self._all_emulators
+            if issubclass(emulator, GaussianProcessEmulator)
+        ]
+
+        self._emulator_registry = {
+            em_cls.model_name().lower(): em_cls for em_cls in self._all_emulators
+        }
+        self._emulator_registry_short_name = {
+            em_cls.short_name(): em_cls for em_cls in self._all_emulators
+        }
+
+    def register_model(self, model_cls: type[Emulator]) -> type[Emulator]:
+        """
+        Register a new emulator model to the registry.
+
[…]
