### 2025-10-20 3b354c5
Update GP subclass registry
- Remove previous module-level registry and use global registry instead
- Fix pickling issue that affects subclass construction outside of the AutoEmulate package
