### 2025-10-20 2da4224
Fix overwrite with default GP subclasses
