### 2025-10-20 893d55b
Update register to handle having args as decorator
