### 2025-10-20 ae975c5
Fix emulator list handling
