### 2025-10-20 25174d1
Refactor compare to use registry directly
