### 2025-10-20 c96732e
ghrs: report 10-20-2304-E839 for alan-turing-institute/autoemulate
