### 2025-10-21 d003870
Refactors and update docstrings
