### 2025-10-21 9ab63ee
Add tests for get_emulator_cls
