### 2025-10-22 b0e4cbe
Update config and add normalization
