+++ autoemulate/experimental/configs/turbulent_radiative_layer_2d.yaml
@@ -15,8 +15,9 @@ formatter_type: "default_channels_first"
 model_params:
   modes1: 16
   modes2: 16
-  width: 20
-  n_blocks: 4
+  hidden_channels: 128
+  # width: 20
+  # n_blocks: 4
 
 # Data configuration - The Well native dataset
 # Setting well_dataset_name automatically uses WELL_NATIVE source type
@@ -26,6 +27,7 @@ data:
   n_steps_input: 4
   n_steps_output: 1
   batch_size: 4
+  use_normalization: true
 
 # No simulator section needed for native Well datasets!
 
