+++ autoemulate/experimental/data/spatiotemporal_dataset.py
@@ -8,6 +8,133 @@ from the_well.data.datasets import WellMetadata
 from torch.utils.data import DataLoader, Dataset
 
 
+class SimpleZScoreNormalization:
+    """Simple Z-score normalization for AutoEmulate datasets.
+
+    Computes mean and std from the data and applies (x - mean) / std normalization.
+    Compatible with The Well's Trainer interface.
+    """
+
+    def __init__(self, data: torch.Tensor, min_std: float = 1e-6):
+        """Initialize normalization with data statistics.
+
+        Parameters
+        ----------
+        data : torch.Tensor
+            Training data tensor of shape [N, T, W, H, C]
+        min_std : float
+            Minimum standard deviation to avoid division by zero
+        """
+        # Compute statistics over all dimensions except channels
+        # Shape: [C] - one mean/std per channel
+        self.mean = data.mean(dim=(0, 1, 2, 3))
+        self.std = data.std(dim=(0, 1, 2, 3)).clamp(min=min_std)
+        self.min_std = min_std
+
+    def normalize(self, x: torch.Tensor) -> torch.Tensor:
+        """Normalize input tensor.
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            Input tensor of shape [..., C]
+
+        Returns
+        -------
+        torch.Tensor
+            Normalized tensor
+        """
+        return (x - self.mean.to(x.device)) / self.std.to(x.device)
+
+    def denormalize(self, x: torch.Tensor) -> torch.Tensor:
+        """Denormalize input tensor.
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            Normalized tensor of shape [..., C]
+
+        Returns
+        -------
+        torch.Tensor
+            Denormalized tensor
+        """
+        return x * self.std.to(x.device) + self.mean.to(x.device)
+
+    def normalize_flattened(
+        self,
+        x: torch.Tensor,
+        mode: str = "variable",  # noqa: ARG002
+    ) -> torch.Tensor:
+        """Normalize tensor where fields are flattened as channels.
+
+        Compatible with The Well's Trainer interface.
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            Input tensor of shape [..., C]
+        mode : str
+            Mode for normalization ("variable" or "constant").
+            Currently both use same stats.
+
+        Returns
+        -------
+        torch.Tensor
+            Normalized tensor
+        """
+        return self.normalize(x)
+
+    def denormalize_flattened(
+        self,
+        x: torch.Tensor,
+        mode: str = "variable",  # noqa: ARG002
+    ) -> torch.Tensor:
+        """Denormalize tensor where fields are flattened as channels.
+
+        Compatible with The Well's Trainer interface.
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            Normalized tensor of shape [..., C]
+        mode : str
+            Mode for denormalization ("variable" or "constant").
+            Currently both use same stats.
+
+        Returns
+        -------
+        torch.Tensor
+            Denormalized tensor
+        """
+        return self.denormalize(x)
+
+    def delta_denormalize_flattened(
+        self,
+        x: torch.Tensor,
+        mode: str = "variable",  # noqa: ARG002
+    ) -> torch.Tensor:
+        """Delta denormalize tensor (for delta models).
+
+        For simple Z-score normalization, delta denormalization only
+        requires std scaling (no mean shift needed).
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            Normalized delta tensor of shape [..., C]
+        mode : str
+            Mode for denormalization ("variable" or "constant").
+            Currently both use same stats.
+
+        Returns
+        -------
+        torch.Tensor
+            Denormalized delta tensor
+        """
[…]
