+++ autoemulate/experimental/configs/turbulent_radiative_layer_2d.yaml
@@ -33,13 +33,13 @@ data:
 
 # Trainer configuration
 trainer:
-  epochs: 100
+  epochs: 200
   device: "cuda"
   
-  optimizer_type: "adam"
+  optimizer_type: "adamw"
   optimizer_params:
-    lr: 0.001
-    weight_decay: 0.0
+    lr: 0.01
+    weight_decay: 0.0001
   
   lr_scheduler_type: "step_lr"
   lr_scheduler_params:
@@ -49,7 +49,7 @@ trainer:
   loss_fn: "vrmse"
   
   val_frequency: 5
-  checkpoint_frequency: 10
+  checkpoint_frequency: 20
   rollout_val_frequency: 10
   max_rollout_steps: 100
   short_validation_length: 10
