### 2025-10-22 809f92b
Update advection diffusion
