### 2025-10-22 a292ea7
Update reaction-diffusion config
