### 2025-10-22 7b8d191
Update turbulent_layer_2d config
