### 2025-10-22 e537535
Update advection-diffusion config
