+++ autoemulate/experimental/configs/advection_diffusion_generated.yaml
@@ -53,7 +53,7 @@ data:
 
 # Trainer configuration
 trainer:
-  epochs: 20
+  epochs: 100
   device: "cuda"           # Use "cuda" for NVIDIA GPU, "mps" for Apple Silicon, "cpu" for CPU
   
   optimizer_type: "adamw"
@@ -69,7 +69,7 @@ trainer:
   loss_fn: "vrmse"
   
   val_frequency: 5
-  checkpoint_frequency: 10
+  checkpoint_frequency: 20
   rollout_val_frequency: 10
   max_rollout_steps: 100
   short_validation_length: 10
@@ -96,10 +96,10 @@ trainer:
 
 # Paths
 paths:
-  output_dir: "./outputs/2025-10-17/advection_diffusion_generated"
+  output_dir: "./outputs/2025-10-22/advection_diffusion_generated"
   model_save_path: "./outputs/advection_diffusion_generated/artifacts/final_model.pt"
   # Optionally save generated data for reuse
-  data_save_path: "./outputs/2025-10-17/data/advection_diffusion_generated"
+  data_save_path: "./outputs/2025-10-22/data/advection_diffusion_generated"
   save_format: "h5"
 
 # Logging
