### 2025-10-22 37e29e2
Add normalization to config
