### 2025-10-22 e9443b8
Add physicsnemo dep
