### 2025-10-23 7c05589
Add to docstring explaining first metric in list is used to pick best model
