### 2025-10-23 a316f54
Remove uneeded attirbutes
