### 2025-10-23 714bc32
Apply suggestions from code review
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
