### 2025-10-23 eb63932
Fix metric import errors
