### 2025-10-23 5f71801
cleaned up evaluate function
