### 2025-10-23 49898af
Refactor to registry submodule
