+++ autoemulate/emulators/__init__.py
@@ -1,9 +1,4 @@
-from collections.abc import Callable
-from typing import overload
-
-import torch
-
-from .base import Emulator, GaussianProcessEmulator
+from .base import Emulator
 from .ensemble import EnsembleMLP, EnsembleMLPDropout
 from .gaussian_process.exact import (
     GaussianProcessCorrelatedMatern32,
@@ -16,207 +11,9 @@ from .nn.mlp import MLP
 from .polynomials import PolynomialRegression
 from .radial_basis_functions import RadialBasisFunctions
 from .random_forest import RandomForest
+from .registry import Registry, _default_registry, get_emulator_class, register
 from .svm import SupportVectorMachine
 from .transformed.base import TransformedEmulator
-
-
-class Registry:
-    """Registry for managing emulators.
-
-    The Registry class maintains collections of emulator classes organized by
-    their properties (e.g., Gaussian Process emulators, PyTorch-based emulators).
-    It provides methods to register new emulators and retrieve them by name.
-    """
-
-    def __init__(self):
-        # Initialize the registry with default emulators, this is not updated
-        self._default_emulators: list[type[Emulator]] = [
-            GaussianProcessMatern32,
-            GaussianProcessRBF,
-            RadialBasisFunctions,
-            PolynomialRegression,
-            MLP,
-            EnsembleMLP,
-        ]
-
-        self._non_pytorch_emulators: list[type[Emulator]] = [
-            LightGBM,
-            SupportVectorMachine,
-            RandomForest,
-        ]
-
-        self._all_emulators: list[type[Emulator]] = [
-            *self._default_emulators,
-            *self._non_pytorch_emulators,
-            GaussianProcessCorrelatedMatern32,
-            GaussianProcessCorrelatedRBF,
-            EnsembleMLPDropout,
-        ]
-
-        self._pytorch_emulators: list[type[Emulator]] = [
-            emulator
-            for emulator in self._all_emulators
-            if emulator not in self._non_pytorch_emulators
-        ]
-        self._gaussian_process_emulators: list[type[Emulator]] = [
-            emulator
-            for emulator in self._all_emulators
-            if issubclass(emulator, GaussianProcessEmulator)
-        ]
-
-        self._emulator_registry = {
-            em_cls.model_name().lower(): em_cls for em_cls in self._all_emulators
-        }
-        self._emulator_registry_short_name = {
-            em_cls.short_name(): em_cls for em_cls in self._all_emulators
-        }
-
-    def register_model(
-        self, model_cls: type[Emulator], overwrite: bool = False
-    ) -> type[Emulator]:
-        """Register a new emulator model to the registry.
-
-        Parameters
-        ----------
-        model_cls: type[Emulator]
-            The emulator class to register.
-        overwrite: bool
-            If True, allows overwriting an existing model with the same name. If False,
-            raises an error if a model with the same name already exists. Defaults to
-            False.
-
-        Returns
-        -------
-        type[Emulator]
-            The registered emulator class (unchanged).
-
-        Raises
-        ------
-        ValueError
-            If overwrite is False and a model with the same name already exists.
-
-        """
-        model_name = model_cls.model_name().lower()
-        short_name = model_cls.short_name()
-
-        # Check if model already exists
-        existing_cls_by_name = self._emulator_registry.get(model_name)
-        existing_cls_by_short_name = self._emulator_registry_short_name.get(short_name)
-
-        if not overwrite:
-            if existing_cls_by_name is not None:
-                raise ValueError(
-                    f"Model with name '{model_name}' already exists. Set overwrite=True"
-                    f" to replace it."
-                )
[…]
