+++ _sources/tutorials/advanced/index.md
@@ -1,0 +1,3 @@
+# Advanced usage
+
+This section covers some of the more advanced features of AutoEmulate, including adding custom emulators.
