### 2025-10-24 d79ff9a
ghrs: report 10-24-2304-BDF4 for alan-turing-institute/autoemulate
