### 2025-10-25 39883d5
ghrs: report 10-25-2304-4FCD for alan-turing-institute/autoemulate
