### 2025-10-26 b7cccf6
ghrs: report 10-26-2304-6D1E for alan-turing-institute/autoemulate
