### 2025-10-27 085b6ca
ghrs: report 10-27-2304-C2B8 for alan-turing-institute/autoemulate
