### 2025-10-28 0c36b4f
Add workspace pyproject
