+++ pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "autox"
+version = "0.0.1"
+requires-python = ">=3.10,<3.13"
+dependencies = []
+
+[tool.uv.sources]
+bird-feeder = { workspace = true }
+
+[tool.uv.workspace]
+members = ["autox/*"]
+# exclude = ["autox/seeds"]
