### 2025-10-28 36f2d97
Init autocast
