+++ autox/autocast/pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "autocast"
+version = "0.1.0"
+description = "AutoCast"
+readme = "README.md"
+authors = []
+requires-python = ">=3.10, <3.13"
+dependencies = []
+
+[build-system]
+requires = ["uv_build>=0.8.15,<0.9.0"]
+build-backend = "uv_build"
