### 2025-10-28 2ba3a75
ghrs: report 10-28-2304-925A for alan-turing-institute/autoemulate
