### 2025-10-29 cc57d95
Add instructions for venv options with virtual workspace
