+++ pyproject.toml
@@ -5,8 +5,10 @@ requires-python = ">=3.10,<3.13"
 dependencies = []
 
 [tool.uv.sources]
-bird-feeder = { workspace = true }
+autoemulate = { path = "autox/autoemulate" }
 
+# Whether to have a combined virtual workspace with single lock or a lock per package
 [tool.uv.workspace]
 members = ["autox/*"]
-# exclude = ["autox/seeds"]
+# Can just exclude a single package from the combined workspace too
+exclude = ["autox/autoemulate"]
