+++ pyproject.toml
@@ -40,6 +40,7 @@ dependencies = [
     "torchrbf>=0.0.1",
     "arviz>=0.21.0",
     "getdist>=1.7.2",
+    "einops>=0.8.1",
 ]
 
 [project.urls]
