### 2025-10-30 efdd711
Refactor and fix shape handling
