+++ autoemulate/core/metrics.py
@@ -86,15 +86,23 @@ class ProbabilisticMetric(Metric):
         """Calculate metric."""
 
 
-class CRPS(ProbabilisticMetric):
+class CRPSMetric(ProbabilisticMetric):
     """Continuous Ranked Probability Score (CRPS) metric.
 
-    Parameters
+    CRPS is a scoring rule for evaluating probabilistic predictions. It reduces to mean
+    absolute error (MAE) for deterministic predictions and generalizes to distributions
+    by measuring the integral difference between predicted and actual CDFs.
+
+    The metric aggregates over batch and target dimensions by computing the mean
+    CRPS across all scalar outputs, making it comparable across different batch
+    sizes and output dimensions.
+
+    Attributes
     ----------
-    name : str
+    name: str
         Display name for the metric.
-    maximize : bool
-        Whether higher values are better. Defaults to False.
+    maximize: bool
+        Whether higher values are better. False for CRPS (lower is better).
     """
 
     name: str = "crps"
@@ -105,69 +113,91 @@ class CRPS(ProbabilisticMetric):
     ) -> TensorLike:
         """Calculate CRPS metric.
 
-        The metric can handle both deterministic predictions (tensors) and probabilistic
-        predictions.
+        The metric handles both deterministic predictions (tensors) and probabilistic
+        predictions (tensors of samples or distributions).
 
-        Aggregation across batch and target dimensions is performed by flattening such
-        that the sum of scores is taken across all samples for each point.
+        Aggregation across batch and target dimensions is performed by computing the
+        mean CRPS across all scalar outputs. This makes the metric comparable across
+        different batch sizes and target dimensions.
 
         Parameters
         ----------
         y_pred: OutputLike
-            Predicted outputs. Can be a tensor or a distribution. If `y_pred` is a
-            tensor of shape (batch_size, *(target_shape)), it is treated as
-            a deterministic prediction and reduces the metric calculation to mean
-            absolute error.
-            If `y_pred` is a tensor of shape
-            `(batch_size, *(target_shape),  n_samples)`, it is treated as a
-            probabilistic prediction and the metric is computed across the samples.
-            If `y_pred` is a distribution, then `n_samples` are drawn from the predicted
-            distribution to estimate the CRPS.
+            Predicted outputs. Can be a tensor or a distribution.
+            - If tensor with shape `(batch_size, *target_shape)`: treated as
+            deterministic prediction (reduces to MAE).
+            - If tensor with shape `(batch_size, *target_shape, n_samples)`: treated as
+            samples from a probabilistic prediction.
+            - If distribution: `n_samples` are drawn to estimate CRPS.
         y_true: TensorLike
-            True target values.
+            True target values of shape `(batch_size, *target_shape)`.
         n_samples: int
             Number of samples to draw from the predicted distribution if `y_pred` is a
             distribution. Defaults to 1000.
 
+        Returns
+        -------
+        TensorLike
+            Mean CRPS score across all batch elements and target dimensions.
+
+        Raises
+        ------
+        ValueError
+            If input types or shapes are incompatible.
         """
         if not isinstance(y_true, TensorLike):
-            raise ValueError(f"Metric not implemented for y_true ({type(y_true)})")
+            raise ValueError(f"y_true must be a tensor, got {type(y_true)}")
 
+        # Ensure 2D y_true for consistent handling
+        y_true = y_true.unsqueeze(-1) if y_true.ndim == 1 else y_true
+
+        # Initialize CRPS metric (computes mean by default)
         crps_metric = ContinuousRankedProbabilityScore()
[…]
