### 2025-10-30 e641aa5
Add to available metrics
