### 2025-10-30 ce05345
add option to plot mean with CI instead of error bars
