### 2025-10-30 2aa650d
add plot of predictions
