### 2025-10-30 f25fe33
Add tests for CRPS
