+++ tests/core/test_metrics.py
@@ -7,15 +7,18 @@ import torch
 import torchmetrics
 from autoemulate.core.metrics import (
     AVAILABLE_METRICS,
+    CRPS,
     MAE,
     MSE,
     R2,
     RMSE,
+    CRPSMetric,
     Metric,
     TorchMetrics,
     get_metric_config,
     get_metric_configs,
 )
+from torch.distributions import Normal
 
 # Tests for the base Metric class
 
@@ -28,7 +31,7 @@ def test_metric_repr():
     metric.maximize = True
 
     repr_str = repr(metric)
-    assert "MetricConfig" in repr_str
+    assert "Metric" in repr_str
     assert "test_metric" in repr_str
     assert "True" in repr_str
 
@@ -326,14 +329,15 @@ def test_get_metric_configs_with_torchmetrics_instances():
 
 def test_get_metric_configs_case_insensitive():
     """Test get_metric_configs is case insensitive for strings."""
-    metrics = ["R2", "RMSE", "mse", "MaE"]
+    metrics = ["R2", "RMSE", "mse", "MaE", "Crps"]
     configs = get_metric_configs(metrics)
 
-    assert len(configs) == 4
+    assert len(configs) == 5
     assert configs[0] == R2
     assert configs[1] == RMSE
     assert configs[2] == MSE
     assert configs[3] == MAE
+    assert configs[4] == CRPS
 
 
 # Integration tests for metrics with actual computation
@@ -408,3 +412,119 @@ def test_metric_configs_workflow():
     assert "rmse" in results
     assert torch.isclose(results["r2"], torch.tensor(1.0))  # Perfect R2
     assert torch.isclose(results["rmse"], torch.tensor(0.0))  # Perfect RMSE
+
+
+# Tests for CRPS metric
+
+
+def test_crps_in_available_metrics():
+    """Test CRPS is in AVAILABLE_METRICS."""
+    assert "crps" in AVAILABLE_METRICS
+    assert AVAILABLE_METRICS["crps"] == CRPS
+
+
+def test_crps_deterministic_reduces_to_mae():
+    """Test CRPS with deterministic predictions reduces to MAE."""
+    y_pred = torch.tensor([[1.0, 2.0], [3.0, 4.0]])
+    y_true = torch.tensor([[1.5, 2.5], [3.5, 4.5]])
+
+    crps_result = CRPS(y_pred, y_true)
+
+    mae = torch.nn.functional.l1_loss(y_pred, y_true)
+    assert torch.isclose(crps_result, mae, rtol=1e-4), "CRPS should equal MAE"
+
+
+@pytest.mark.parametrize(
+    ("batch_size", "n_targets", "n_samples"),
+    [
+        (10, 1, 5),  # Single target
+        (10, 3, 5),  # Multiple targets
+        (5, 2, 10),  # More samples than batch
+        (100, 5, 20),  # Large batch
+        (1, 1, 100),  # Single sample, many ensemble members
+    ],
+)
+def test_crps_with_various_shapes(batch_size, n_targets, n_samples):
+    """Test CRPS with various tensor shapes."""
+    y_pred = torch.randn(batch_size, n_targets, n_samples)
+    y_true = torch.randn(batch_size, n_targets)
+
+    result = CRPS(y_pred, y_true)
+
+    assert result.ndim == 0, "Result should be a scalar tensor"
+    assert isinstance(result, torch.Tensor)
+    assert result >= 0, "CRPS should be non-negative"
+
+
+def test_crps_with_distribution():
+    """Test CRPS with distribution input."""
+    batch_size, n_targets = 10, 3
+    y_true = torch.randn(batch_size, n_targets)
+
+    # Create a distribution
+    mean = torch.randn(batch_size, n_targets)
[…]
