### 2025-10-30 01cd53e
add option to include x/y labels in plots
