+++ autoemulate/core/compare.py
@@ -591,6 +591,8 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
     def plot(  # noqa: PLR0912, PLR0915
         self,
         model_obj: int | Emulator | Result,
+        input_names: list[str] | None = None,
+        output_names: list[str] | None = None,
         input_index: list[int] | int | None = None,
         output_index: list[int] | int | None = None,
         input_ranges: dict | None = None,
@@ -608,6 +610,10 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
         model_obj: int | Emulator | Result
             The model to plot. Can be an integer ID of a Result, an Emulator instance,
             or a Result instance.
+        input_names: list[str] | None
+            The names of the input features. If None, generic names are used.
+        output_names: list[str] | None
+            The names of the output features. If None, generic names are used.
         input_index: int
             The index of the input feature to plot against the output.
         output_index: int
@@ -702,6 +708,26 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
         fig, axs = plt.subplots(nrows, ncols, figsize=figsize, squeeze=False)
         axs = axs.flatten()
 
+        if input_names is not None:
+            if len(input_names) != n_features:
+                msg = (
+                    "Length of input_names does not match number of input features. "
+                    f"Expected {n_features}, got {len(input_names)}."
+                )
+                raise ValueError(msg)
+        else:
+            input_names = [f"$x_{i}$" for i in range(n_features)]
+
+        if output_names is not None:
+            if len(output_names) != n_outputs:
+                msg = (
+                    "Length of output_names does not match number of outputs. "
+                    f"Expected {n_outputs}, got {len(output_names)}."
+                )
+                raise ValueError(msg)
+        else:
+            output_names = [f"$y_{i}$" for i in range(n_outputs)]
+
         plot_index = 0
         for out_idx in output_index:
             for in_idx in input_index:
@@ -752,9 +778,9 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
                         y_pred_subset[:, out_idx],
                         y_variance[:, out_idx] if y_variance is not None else None,
                         ax=axs[plot_index],
-                        title=f"$x_{in_idx}$ vs. $y_{out_idx}$",
-                        input_index=in_idx,
-                        output_index=out_idx,
+                        title=f"{input_names[in_idx]} vs. {output_names[out_idx]}",
+                        input_label=input_names[in_idx],
+                        output_label=output_names[out_idx],
                         r2_score=r2_score,
                         error_style=error_style,
                     )
@@ -773,6 +799,7 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
     def plot_preds(
         self,
         model_obj: int | Emulator | Result,
+        output_names: list[str] | None = None,
         figsize=None,
         ncols: int = 3,
         fname: str | None = None,
@@ -785,6 +812,9 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
         model_obj: int | Emulator | Result
             The model to plot. Can be an integer ID of a Result, an Emulator instance,
             or a Result instance.
+        output_names: list[str] | None
+            The names of the outputs to use in the plot titles. If None, generic names
+            like "y_0", "y_1", etc. are used.
         figsize: tuple[int, int] | None
             The size of the figure to create. If None, it is set based on the number
             of outputs.
@@ -825,6 +855,16 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
         fig, axs = plt.subplots(nrows, ncols, figsize=figsize, squeeze=False)
[…]
