### 2025-10-30 6a22bad
ensure all arrays are 2D
