### 2025-10-30 d644cfa
Update `Metric` and `evaluate()` API to support `OutputLike`
