+++ autoemulate/core/compare.py
@@ -12,11 +12,7 @@ from torch.distributions import Transform
 
 from autoemulate.core.device import TorchDeviceMixin
 from autoemulate.core.logging_config import get_configured_logger
-from autoemulate.core.metrics import (
-    TorchMetrics,
-    get_metric_config,
-    get_metric_configs,
-)
+from autoemulate.core.metrics import TorchMetrics, get_metric_config, get_metric_configs
 from autoemulate.core.model_selection import bootstrap, evaluate
 from autoemulate.core.plotting import (
     calculate_subplot_layout,
