+++ pyproject.toml
@@ -16,18 +16,15 @@ classifiers = [
     "Topic :: Scientific/Engineering",
 ]
 dependencies = [
-    "mogp-emulator>=0.7.2",
     "matplotlib>=3.7.2",
     "seaborn>=0.12.2",
     "scikit-learn>=1.3.0,<1.6.0",
     "pandas>=2.1",
     "torch>=2.1.0",
-    "skorch>=0.15.0",
     "scipy>=1.11.3",
     "numpy>=1.24",
     "joblib>=1.3.2",
     "tqdm>=4.66.2",
-    "iprogress>=0.4",
     "lightgbm>=4.3.0",
     "ipywidgets>=8.1.2",
     "gpytorch>=1.12",
@@ -36,7 +33,6 @@ dependencies = [
     "anytree>=2.12.1",
     "torchmetrics>=1.7.1",
     "pyro-ppl>=1.9.1",
-    "corner>=2.2.3",
     "torchrbf>=0.0.1",
     "arviz>=0.21.0",
     "getdist>=1.7.2",
