### 2025-10-30 5a6e48d
reduce word count
