### 2025-10-31 f54a16c
Update autoemulate/core/compare.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
