### 2025-10-31 303fb19
Update autoemulate/core/plotting.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
