+++ autoemulate/core/plotting.py
@@ -100,7 +100,7 @@ def plot_xy(
         if error_style.lower() not in ["bars", "fill"]:
             msg = "error_style must be one of ['bars', 'fill']"
             raise ValueError(msg)
-        if error_style == "bars":
+        if error_style.lower() == "bars":
             ax.errorbar(
                 x_sorted,
                 y_pred_sorted,
