### 2025-10-31 b6aee64
change CI to 2SD
