### 2025-10-31 7b17732
Add R2 as an argument
