### 2025-10-31 db288c5
Update `.predict_mean()` to `.predict()`
