### 2025-10-31 638b446
Update API for getting metric config
