### 2025-10-31 0559b9c
Update paper/paper.md
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
