### 2025-10-31 7e9a288
Fix pyright lints, update test, fix for 1D
