### 2025-10-31 f43dbb9
ghrs: report 10-31-2304-4C94 for alan-turing-institute/autoemulate
