### 2025-11-01 bff707f
ghrs: report 11-01-2304-D8C0 for alan-turing-institute/autoemulate
