### 2025-11-02 afb3387
ghrs: report 11-02-2304-AE6A for alan-turing-institute/autoemulate
