### 2025-11-03 0c4e7c2
Revise fn name to remove config
