+++ tests/core/test_metrics.py
@@ -15,8 +15,8 @@ from autoemulate.core.metrics import (
     CRPSMetric,
     Metric,
     TorchMetrics,
-    get_metric_config,
-    get_metric_configs,
+    get_metric,
+    get_metrics,
 )
 from torch.distributions import Normal
 
@@ -176,7 +176,7 @@ def test_mae_computation():
 
 def test_get_metric_config_with_string_r2():
     """Test get_metric_config with 'r2' string."""
-    config = get_metric_config("r2")
+    config = get_metric("r2")
 
     assert config == R2
     assert config.name == "r2"
@@ -185,7 +185,7 @@ def test_get_metric_config_with_string_r2():
 
 def test_get_metric_config_with_string_rmse():
     """Test get_metric_config with 'rmse' string."""
-    config = get_metric_config("rmse")
+    config = get_metric("rmse")
 
     assert config == RMSE
     assert config.name == "rmse"
@@ -194,7 +194,7 @@ def test_get_metric_config_with_string_rmse():
 
 def test_get_metric_config_with_string_mse():
     """Test get_metric_config with 'mse' string."""
-    config = get_metric_config("mse")
+    config = get_metric("mse")
 
     assert config == MSE
     assert config.name == "mse"
@@ -203,7 +203,7 @@ def test_get_metric_config_with_string_mse():
 
 def test_get_metric_config_with_string_mae():
     """Test get_metric_config with 'mae' string."""
-    config = get_metric_config("mae")
+    config = get_metric("mae")
 
     assert config == MAE
     assert config.name == "mae"
@@ -212,9 +212,9 @@ def test_get_metric_config_with_string_mae():
 
 def test_get_metric_config_case_insensitive():
     """Test get_metric_config is case insensitive."""
-    config_upper = get_metric_config("R2")
-    config_lower = get_metric_config("r2")
-    config_mixed = get_metric_config("R2")
+    config_upper = get_metric("R2")
+    config_lower = get_metric("r2")
+    config_mixed = get_metric("R2")
 
     assert config_upper == config_lower == config_mixed == R2
 
@@ -225,7 +225,7 @@ def test_get_metric_config_with_torchmetrics_instance():
         metric=torchmetrics.R2Score, name="custom_r2", maximize=True
     )
 
-    config = get_metric_config(custom_metric)
+    config = get_metric(custom_metric)
 
     assert config == custom_metric
     assert config.name == "custom_r2"
@@ -234,7 +234,7 @@ def test_get_metric_config_with_torchmetrics_instance():
 def test_get_metric_config_invalid_string():
     """Test get_metric_config with invalid string raises ValueError."""
     with pytest.raises(ValueError, match="Unknown metric shortcut") as excinfo:
-        get_metric_config("invalid_metric")
+        get_metric("invalid_metric")
 
     assert "Unknown metric shortcut" in str(excinfo.value)
     assert "invalid_metric" in str(excinfo.value)
@@ -244,7 +244,7 @@ def test_get_metric_config_invalid_string():
 def test_get_metric_config_unsupported_type():
     """Test get_metric_config with unsupported type raises ValueError."""
     with pytest.raises(ValueError, match="Unsupported metric type") as excinfo:
-        get_metric_config(123)  # type: ignore[arg-type]
+        get_metric(123)  # type: ignore[arg-type]
 
     assert "Unsupported metric type" in str(excinfo.value)
 
@@ -252,7 +252,7 @@ def test_get_metric_config_unsupported_type():
 def test_get_metric_config_with_none():
     """Test get_metric_config with None raises ValueError."""
[…]
