### 2025-11-03 66d0677
Fix docstring style
