### 2025-11-03 6b99c47
Update to store metrics by class
