### 2025-11-03 922dff1
Update base class with ordering
