### 2025-11-03 f50b905
Add __str__  alongside __repr__
