### 2025-11-03 4182fe8
Additional API updates, add tests, fix tests
