+++ tests/core/test_compare.py
@@ -5,6 +5,8 @@ import pytest
 import torch
 from autoemulate.core.compare import AutoEmulate
 from autoemulate.core.device import SUPPORTED_DEVICES, check_torch_device_is_available
+from autoemulate.core.metrics import Metric, get_metric
+from autoemulate.core.types import OutputLike, TensorLike
 from autoemulate.emulators import DEFAULT_EMULATORS
 from autoemulate.emulators.base import Emulator
 from torch.distributions import Transform
@@ -187,11 +189,12 @@ def test_ae_with_different_evaluation_metrics(
     # Verify that all specified metrics are in test_metrics for each result
     for result in ae.results:
         assert result.test_metrics is not None
+        metric_names = [m.name for m in result.test_metrics]
         for metric_name in evaluation_metrics:
-            assert metric_name in result.test_metrics
+            assert metric_name in metric_names
             # Verify the metric value is a tuple of (mean, std)
-            assert isinstance(result.test_metrics[metric_name], tuple)
-            assert len(result.test_metrics[metric_name]) == 2
+            assert isinstance(result.test_metrics[get_metric(metric_name)], tuple)
+            assert len(result.test_metrics[get_metric(metric_name)]) == 2
 
 
 @pytest.mark.parametrize(
@@ -235,7 +238,8 @@ def test_ae_with_tuning_and_evaluation_metric_combinations(
     for result in ae.results:
         assert result.test_metrics is not None
         for metric_name in evaluation_metrics:
-            assert metric_name in result.test_metrics
+            metric_names = [m.name for m in result.test_metrics]
+            assert metric_name in metric_names
 
 
 def test_ae_with_same_tuning_and_evaluation_metric(sample_data_for_ae_compare):
@@ -260,7 +264,8 @@ def test_ae_with_same_tuning_and_evaluation_metric(sample_data_for_ae_compare):
     assert ae.evaluation_metrics[0].name == "rmse"
 
     for result in ae.results:
-        assert "rmse" in result.test_metrics
+        metric_names = [m.name for m in result.test_metrics]
+        assert "rmse" in metric_names
 
 
 def test_ae_with_maximizing_and_minimizing_metrics(sample_data_for_ae_compare):
@@ -290,7 +295,134 @@ def test_ae_with_maximizing_and_minimizing_metrics(sample_data_for_ae_compare):
 
     # Verify all metrics are computed
     for result in ae.results:
+        metric_names = [m.name for m in result.test_metrics]
         assert len(result.test_metrics) == 4
-        assert all(
-            metric in result.test_metrics for metric in ["r2", "rmse", "mse", "mae"]
+        assert all(metric in metric_names for metric in ["r2", "rmse", "mse", "mae"])
+
+
+def test_ae_with_custom_evaluation_metrics(sample_data_for_ae_compare):
+    """Test AutoEmulate with custom metric implementations alongside built-in
+    metrics."""
+
+    class CustomMSEMetric(Metric):
+        """Custom MSE metric for testing custom metric functionality."""
+
+        name = "custom_mse"
+        maximize = False
+
+        def __call__(
+            self,
+            y_pred: OutputLike,
+            y_true: TensorLike,
+            n_samples: int = 1000,  # noqa: ARG002
+        ) -> TensorLike:
+            """Calculate mean squared error."""
+            assert isinstance(y_pred, torch.Tensor)
+            return torch.mean((y_pred - y_true) ** 2)
+
+    custom_mse = CustomMSEMetric()
+    x, y = sample_data_for_ae_compare
+    models: list[str | type[Emulator]] = ["mlp", "RandomForest"]
+
[…]
