### 2025-11-03 ba41894
Enable tuning in tests with tuning metrics
