### 2025-11-03 c794578
Update workflows to use uv
