+++ .github/workflows/precommit.yaml
@@ -30,12 +30,16 @@ jobs:
           - uses: actions/setup-python@v5
             with:
               python-version: "3.12"
+          - name: Install uv
+            uses: astral-sh/setup-uv@v7
+            with:
+              python-version: "3.12"
+              enable-cache: true
+              cache-suffix: precommit
+              activate-environment: true
           - name: Install dependencies
             run: |
-              python -m venv .venv
-              source .venv/bin/activate
-              python -m pip install --upgrade pip
-              pip install -e .[dev,spatiotemporal]
+              uv pip install -e .[dev,spatiotemporal]
 
           - uses: pre-commit/action@v3.0.0
             with:
