+++ .github/workflows/docs.yaml
@@ -24,13 +24,18 @@ jobs:
             - uses: actions/setup-python@v5
               with:
                 python-version: 3.12
+            - name: Install uv
+              uses: astral-sh/setup-uv@v7
+              with:
+                python-version: 3.12
+                enable-cache: true
+                cache-suffix: docs
+                activate-environment: true
             - uses: nikeee/setup-pandoc@v1
-            - name: "Upgrade pip"
-              run: pip install --upgrade pip
             - name: "Install jupyterbook"
-              run: pip install -r docs/requirements.txt
+              run: uv pip install -r docs/requirements.txt
             - name: "Install autoemulate"
-              run: pip install -e .
+              run: uv pip install -e .
             - name: "Run jupyterbook"
               run: jupyter-book build docs --all 
             - name: "Deploy"
