+++ .github/workflows/release.yaml
@@ -10,17 +10,24 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       - name: Set up Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: '3.10'
 
+      - name: Install uv
+        uses: astral-sh/setup-uv@v7
+        with:
+          python-version: '3.10'
+          enable-cache: true
+          cache-suffix: release
+          activate-environment: true
+
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          pip install build twine
+          uv pip install build twine
 
       - name: Build package
         run: python -m build
