### 2025-11-03 fb95fe1
Update pre-commit workflow for uv
