### 2025-11-03 f2a7eb8
Update jupyter-book dep to <2.0.0
