+++ pyproject.toml
@@ -61,7 +61,7 @@ dev = [
     "sphinx-autodoc-typehints>=1.24.0",
     "black>=23.10.1",
     "pre-commit>=3.5.0",
-    "jupyter-book>=1.0.0",
+    "jupyter-book>=1.0.0,<2.0.0",
     "pytest-cov>=4.1.0",
     "coverage>=7.6.4",
     "plotnine>=0.13.6",
