### 2025-11-03 9689f5d
Remove obsolete requirements
