+++ docs/requirements.txt
@@ -1,4 +1,0 @@
-jupyter-book
-jupyter-cache
-matplotlib
-numpy
