### 2025-11-04 2b52df3
Revise first metric based on results ordering
