### 2025-11-04 b392472
Add validation_data to Emulator API
