### 2025-11-04 c630317
Add output_to_tensor conversion, update usage
