### 2025-11-04 87a2a0b
Add conformal module
