### 2025-11-04 88153bf
Add initial conformal impl and test
