### 2025-11-04 ef684d3
Add ConformalMLP to registry and re-export
