### 2025-11-04 e76ed72
Add support for case when no cal data provided
