### 2025-11-04 340d1e7
Update conformal MLP with kwargs
