### 2025-11-04 b239d69
Updates sample with expected torch.Size, remove GaussianLike case
