### 2025-11-04 699f763
Fix remaining metrics checks with str
