### 2025-11-04 66b76ec
Add MLP kwargs for ensemble versions
