### 2025-11-04 d0ff395
ghrs: report 11-04-2304-5BBF for alan-turing-institute/autoemulate
