+++ autoemulate/emulators/conformal.py
@@ -1,13 +1,77 @@
 import math
+from typing import Literal
 
 import torch
-from torch import Tensor, nn
+from torch import nn
 from torch.optim.lr_scheduler import LRScheduler
 
 from autoemulate.core.device import TorchDeviceMixin
 from autoemulate.core.types import DeviceLike, DistributionLike, TensorLike, TuneParams
 from autoemulate.emulators.base import Emulator, PyTorchBackend
 from autoemulate.emulators.nn.mlp import MLP
+
+
+class QuantileLoss(nn.Module):
+    """Quantile loss for quantile regression.
+
+    This loss function asymmetrically penalizes over- and under-predictions, enabling
+    the model to learn specific quantiles of the conditional distribution.
+    """
+
+    def __init__(self, quantile: float):
+        """Initialize quantile loss.
+
+        Parameters
+        ----------
+        quantile: float
+            Target quantile level in (0, 1). For example, 0.1 for 10th percentile, 0.5
+            for median, 0.9 for 90th percentile.
+        """
+        super().__init__()
+        if not 0 < quantile < 1:
+            msg = f"Quantile must be in (0, 1), got {quantile}"
+            raise ValueError(msg)
+        self.quantile = quantile
+
+    def forward(self, y_pred: TensorLike, y_true: TensorLike) -> TensorLike:
+        """Compute quantile loss.
+
+        Parameters
+        ----------
+        y_pred: TensorLike
+            Predicted values.
+        y_true: TensorLike
+            True target values.
+
+        Returns
+        -------
+        TensorLike
+            Scalar loss value.
+        """
+        errors = y_true - y_pred
+        return torch.max(self.quantile * errors, (self.quantile - 1) * errors).mean()
+
+
+class QuantileMLP(MLP):
+    """MLP with quantile loss for quantile regression."""
+
+    def __init__(self, quantile: float, **kwargs):
+        """Initialize quantile MLP.
+
+        Parameters
+        ----------
+        quantile: float
+            Target quantile level in (0, 1).
+        **kwargs
+            Keyword arguments passed to MLP parent class.
+        """
+        super().__init__(**kwargs)
+        self.quantile = quantile
+        self.quantile_loss = QuantileLoss(quantile)
+
+    def loss_func(self, y_pred, y_true):
+        """Quantile loss function."""
+        return self.quantile_loss(y_pred, y_true)
 
 
 class Conformal(Emulator):
@@ -26,6 +90,8 @@ class Conformal(Emulator):
         device: DeviceLike | None = None,
         calibration_ratio: float = 0.2,
         n_samples: int = 1000,
+        method: Literal["split", "quantile"] = "split",
+        quantile_emulator_kwargs: dict | None = None,
     ):
         """Initialize a conformal emulator.
 
@@ -42,8 +108,16 @@ class Conformal(Emulator):
             Fraction of the training data to reserve for calibration if explicit
             validation data is not provided. Must lie in (0, 1). Defaults to 0.2.
         n_samples: int
-            Number of samples used for sampling-based predictions or
-            internal procedures. Defaults to 1000.
+            Number of samples used for sampling-based predictions or internal
+            procedures. Defaults to 1000.
+        method: Literal["split", "quantile"]
+            Conformalization method to use:
+            - "split": Standard split conformal with constant-width intervals
+            - "quantile": Conformalized Quantile Regression (CQR) with input-dependent
+              intervals. Defaults to "split".
+        quantile_emulator_kwargs: dict | None
+            Additional keyword arguments for the quantile emulators when
+            method="quantile". Defaults to None.
         """
         self.emulator = emulator
[…]
