### 2025-11-05 babfd7e
Fix default dropout_prob value
