### 2025-11-05 5e6caa5
Add neuralprocesses dep
