### 2025-11-05 9366a2e
Add function to generate MLP docstrings
