+++ autoemulate/emulators/ensemble.py
@@ -7,7 +7,7 @@ from torch.optim.lr_scheduler import LRScheduler
 from autoemulate.core.device import TorchDeviceMixin
 from autoemulate.core.types import DeviceLike, GaussianLike, TensorLike, TuneParams
 from autoemulate.emulators.base import DropoutTorchBackend, Emulator, GaussianEmulator
-from autoemulate.emulators.nn.mlp import MLP
+from autoemulate.emulators.nn.mlp import MLP, _generate_mlp_docstring
 from autoemulate.transforms.standardize import StandardizeTransform
 from autoemulate.transforms.utils import make_positive_definite
 
@@ -138,7 +138,6 @@ class EnsembleMLP(Ensemble):
         y: TensorLike,
         standardize_x: bool = True,
         standardize_y: bool = True,
-        n_emulators: int = 4,
         activation_cls: type[nn.Module] = nn.ReLU,
         loss_fn_cls: type[nn.Module] = nn.MSELoss,
         epochs: int = 100,
@@ -154,61 +153,23 @@ class EnsembleMLP(Ensemble):
         device: DeviceLike | None = None,
         scheduler_cls: type[LRScheduler] | None = None,
         scheduler_params: dict | None = None,
+        n_emulators: int = 4,
     ):
-        """
-        Initialize an ensemble of MLPs.
-
-        Parameters
-        ----------
-        x: TensorLike
-            Input data tensor of shape (batch_size, n_features).
-        y: TensorLike
-            Target values tensor of shape (batch_size, n_outputs).
-        standardize_x: bool
-            Whether to standardize the input data. Defaults to True.
-        standardize_y: bool
-            Whether to standardize the output data. Defaults to True.
+        additional_parameters_docstring = """
         n_emulators: int
             Number of MLP emulators to create in the ensemble. Defaults to 4.
-        device: DeviceLike | None
-            Device to run the model on (e.g., "cpu", "cuda"). Defaults to None.
-        activation_cls: type[nn.Module]
-            Activation function to use in the hidden layers. Defaults to `nn.ReLU`.
-        layer_dims: list[int] | None
-            Dimensions of the hidden layers. If None, defaults to [32, 16].
-            Defaults to None.
-        weight_init: str
-            Weight initialization method. Options are "default", "normal", "uniform",
-            "zeros", "ones", "xavier_uniform", "xavier_normal", "kaiming_uniform",
-            "kaiming_normal". Defaults to "default".
-        scale: float
-            Scale parameter for weight initialization methods. Used as:
-            - gain for Xavier methods
-            - std for normal distribution
-            - bound for uniform distribution (range: [-scale, scale])
-            - ignored for Kaiming methods (uses optimal scaling)
-            Defaults to 1.0.
-        bias_init: str
-            Bias initialization method. Options: "zeros", "default":
-                - "zeros" initializes biases to zero
-                - "default" uses PyTorch's default uniform initialization
-        dropout_prob: float | None
-            Dropout probability for regularization. If None, no dropout is applied.
-            Defaults to None.
-        lr: float
-            Learning rate for the optimizer. Defaults to 1e-2.
-        params_size: int
-            Number of parameters to predict per output dimension. Defaults to 1.
-        random_seed: int | None
-            Random seed for reproducibility. If None, no seed is set. Defaults to None.
-        device: DeviceLike | None
-            Device to run the model on (e.g., "cpu", "cuda", "mps"). Defaults to None.
-        scheduler_cls: type[LRScheduler] | None
-            Learning rate scheduler class. If None, no scheduler is used. Defaults to
-            None.
-        scheduler_params: dict | None
-            Additional keyword arguments related to the scheduler.
         """
+        self.__doc__ = f"""
+        Initialize an ensemble of MLPs.
+
+        {
[…]
