### 2025-11-05 3c609fd
Add n_samples to APIs
