+++ autoemulate/core/compare.py
@@ -76,6 +76,7 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
         log_level: str = "progress_bar",
         tuning_metric: str | Metric = "r2",
         evaluation_metrics: list[str | Metric] | None = None,
+        n_samples: int = 1000,
     ):
         """
         Initialize the AutoEmulate class.
@@ -135,6 +136,9 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
             Each entry can be a string shortcut or a MetricConfig object.
             IMPORTANT: The first metric in the list is used to
             determine the best model.
+        n_samples: int
+            Number of samples to generate to predict mean when emulator does not have a
+            mean directly available. Defaults to 1000.
         """
         Results.__init__(self)
         self.random_seed = random_seed
@@ -192,6 +196,7 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
         # Set up logger and ModelSerialiser for saving models
         self.logger, self.progress_bar = get_configured_logger(log_level)
         self.model_serialiser = ModelSerialiser(self.logger)
+        self.n_samples = n_samples
 
         # Run compare
         self.compare()
@@ -489,6 +494,7 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
                                 n_bootstraps=self.n_bootstraps,
                                 device=self.device,
                                 metrics=self.evaluation_metrics,
+                                n_samples=self.n_samples,
                             )
                             test_metrics = bootstrap(
                                 transformed_emulator,
@@ -497,6 +503,7 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin, Results):
                                 n_bootstraps=self.n_bootstraps,
                                 device=self.device,
                                 metrics=self.evaluation_metrics,
+                                n_samples=self.n_samples,
                             )
 
                             # Log all test metrics from test_metrics dictionary
