### 2025-11-05 72d9c95
Added Project email to the readme file
