### 2025-11-05 bb83378
ghrs: report 11-05-2304-ACA3 for alan-turing-institute/autoemulate
