### 2025-11-06 4f19a96
ghrs: report 11-06-2304-3CE6 for alan-turing-institute/autoemulate
