### 2025-11-07 76ec652
Remove obsolete return type
