### 2025-11-07 cbc9821
Update docstring and kwarg order
