### 2025-11-07 edbf1d2
Fix docstrings
