### 2025-11-07 52fa84c
Fix docstring
