### 2025-11-07 dbd9f2d
Revise type hints
