### 2025-11-07 ac897c8
Fix assertion in test
