### 2025-11-07 f2c645e
ghrs: report 11-07-2304-9AA8 for alan-turing-institute/autoemulate
