### 2025-11-08 b92d520
ghrs: report 11-08-2304-BEBA for alan-turing-institute/autoemulate
