### 2025-11-10 baefdf4
reshape y_true and y_train if necessary
