### 2025-11-11 7992e00
add optional metric_kwargs to model selection functions
