### 2025-11-11 ce920a7
pass msll metric_kwargs in compare loop
