### 2025-11-11 deabeec
use global variance of training data to match GPyTorch implementation
