### 2025-11-11 c52e857
add option to average msll over tasks or not
