### 2025-11-11 70f4ca7
add tests for uq aware metrics used with compare
