### 2025-11-11 2d7ff40
ghrs: report 11-11-2304-20CD for alan-turing-institute/autoemulate
