### 2025-11-12 08f79b0
ghrs: report 11-12-2304-36EB for alan-turing-institute/autoemulate
