### 2025-11-13 0728857
Add with_grad to docstring
