### 2025-11-13 ba2ce49
Revise return docstring
