### 2025-11-13 183dc76
Update method str to "constant"
Also revises comments
