### 2025-11-13 2c3c618
Add initial NP with neuralprocesses
Co-authored-by: Radka Jersakova <r.jersakova@gmail.com>
Co-authored-by: Paolo Conti <ContiPaolo@users.noreply.github.com>
Co-authored-by: Christopher Iliffe Sprague <cisprague@users.noreply.github.com>
Co-authored-by: Edwin <EdwinB12@users.noreply.github.com>
Co-authored-by: Marjan Famili <marjanfamili@users.noreply.github.com>
