### 2025-11-13 18c39c1
Update autoemulate/core/metrics.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
