### 2025-11-13 8955fc8
use metric_kwargs when calling evaluate
