### 2025-11-13 e0cf73c
Extend docstring and add reference
