### 2025-11-13 dcbfe72
Update docstring to explain adding new methods
