### 2025-11-13 87ebca3
add MetricParams dataclass
