+++ autoemulate/core/metrics.py
@@ -4,7 +4,9 @@ from __future__ import annotations
 
 from abc import abstractmethod
 from collections.abc import Sequence
+from dataclasses import dataclass
 from functools import partial, total_ordering
+from typing import Literal
 
 import torch
 import torchmetrics
@@ -17,6 +19,30 @@ from autoemulate.core.types import (
     TensorLike,
     TorchMetricsLike,
 )
+
+
+@dataclass
+class MetricParams:
+    """
+    Parameters for metric calculations.
+
+    Attributes
+    ----------
+    n_samples: int
+        Number of samples to draw from the predicted distribution if `y_pred` is a
+        distribution. Defaults to 1000.
+    y_train: TensorLike | None
+        Training target values. In MSLL used to parameterize the trivial model for
+        standardization. If None, mean log loss is computed without standardization.
+    reduction: Literal["mean", "none"]
+        Reduction method to apply to the final metric scores computer per task.
+        Options are 'mean' or 'none'. Defaults to 'mean'.
+    """
+
+    n_samples: int = 1000
+    y_train: TensorLike | None = None
+    reduction: Literal["mean", "none"] = "mean"
+    metric_kwargs: dict | None = None  # supports subclasses with arbitrary new kwargs
 
 
 @total_ordering
@@ -58,7 +84,10 @@ class Metric:
 
     @abstractmethod
     def __call__(
-        self, y_pred: OutputLike, y_true: TensorLike, n_samples: int = 1000
+        self,
+        y_pred: OutputLike,
+        y_true: TensorLike,
+        metric_params: MetricParams | None = None,
     ) -> TensorLike:
         """Calculate metric."""
 
@@ -87,7 +116,10 @@ class TorchMetrics(Metric):
         self.maximize = maximize
 
     def __call__(
-        self, y_pred: OutputLike, y_true: TensorLike, n_samples: int = 1000
+        self,
+        y_pred: OutputLike,
+        y_true: TensorLike,
+        metric_params: MetricParams | None = None,
     ) -> TensorLike:
         """Calculate metric."""
         if not isinstance(y_pred, OutputLike):
@@ -95,12 +127,17 @@ class TorchMetrics(Metric):
         if not isinstance(y_true, TensorLike):
             raise ValueError(f"Metric not implemented for y_true ({type(y_true)})")
 
+        if metric_params is None:
+            metric_params = MetricParams()
+
         # Handle probabilistic predictions
         if isinstance(y_pred, DistributionLike):
             try:
                 y_pred = y_pred.mean
             except Exception:
-                y_pred = y_pred.rsample(torch.Size([n_samples])).mean(dim=0)
+                y_pred = y_pred.rsample(torch.Size([metric_params.n_samples])).mean(
+                    dim=0
+                )
         metric = self.metric()
         metric.to(y_pred.device)
 
@@ -117,7 +154,10 @@ class ProbabilisticMetric(Metric):
 
     @abstractmethod
     def __call__(
-        self, y_pred: OutputLike, y_true: TensorLike, n_samples: int = 1000
+        self,
+        y_pred: OutputLike,
+        y_true: TensorLike,
+        metric_params: MetricParams | None = None,
     ) -> TensorLike:
         """Calculate metric."""
 
@@ -145,7 +185,10 @@ class CRPSMetric(ProbabilisticMetric):
     maximize: bool = False
 
     def __call__(
-        self, y_pred: OutputLike, y_true: TensorLike, n_samples: int = 1000
+        self,
+        y_pred: OutputLike,
+        y_true: TensorLike,
+        metric_params: MetricParams | None = None,
     ) -> TensorLike:
         """Calculate CRPS metric.
 
[…]
