### 2025-11-13 b87c96d
Add conformal MLP subclass factory and subclasses
