+++ autoemulate/emulators/conformal.py
@@ -1,4 +1,5 @@
 import math
+import sys
 from collections.abc import Callable
 from typing import Literal
 
@@ -461,3 +462,217 @@ class ConformalMLP(Conformal, PyTorchBackend):
     def get_tune_params() -> TuneParams:
         """Return a dictionary of hyperparameters to tune."""
         return MLP.get_tune_params()
+
+
+def create_conformal_subclass(
+    name: str,
+    conformal_mlp_base_class: type[ConformalMLP],
+    method: Literal["constant", "quantile"],
+    auto_register: bool = True,
+    overwrite: bool = True,
+    **fixed_kwargs,
+) -> type[ConformalMLP]:
+    """
+    Create a subclass of ConformalMLP with given fixed_kwargs.
+
+    This function creates a subclass of ConformalMLP where certain parameters
+    are fixed to specific values, reducing the parameter space for tuning.
+
+    The created subclass is automatically registered with the main emulator Registry
+    (unless auto_register=False), making it discoverable by AutoEmulate.
+
+    Parameters
+    ----------
+    name: str
+        Name for the created subclass.
+    conformal_mlp_base_class: type[ConformalMLP]
+        Base class to inherit from (typically ConformalMLP).
+    method: Literal["constant", "quantile"]
+        Conformalization method to use in the subclass.
+    auto_register : bool
+        Whether to automatically register the created subclass with the main emulator
+        Registry. Defaults to True.
+    overwrite : bool
+        Whether to allow overwriting an existing class with the same name in the
+        main Registry. Useful for interactive development in notebooks. Defaults to
+        True.
+    **fixed_kwargs
+        Keyword arguments to fix in the subclass. These parameters will be
+        set to the provided values and excluded from hyperparameter tuning.
+
+    Returns
+    -------
+    type[ConformalMLP]
+        A new subclass of ConformalMLP with the specified parameters fixed.
+        The returned class can be pickled and used like any other GP emulator.
+
+    Raises
+    ------
+    ValueError
+        If `name` matches `model_name()` or `short_name()` of an already registered
+        emulator in the main Registry and `overwrite=False`.
+
+    Notes
+    -----
+    - Fixed parameters are automatically excluded from `get_tune_params()` to prevent
+    them from being included in hyperparameter optimization.
+    - Pickling: The created subclass is registered in the caller's module namespace,
+    ensuring it can be pickled and unpickled correctly even when created in downstream
+    code that uses autoemulate as a dependency.
+    - If auto_register=True (default), the class is also added to the main Registry.
+    """
+    standardize_x = fixed_kwargs.get("standardize_x", True)
+    standardize_y = fixed_kwargs.get("standardize_y", True)
+    activation_cls: type[nn.Module] = fixed_kwargs.get("activation_cls", nn.ReLU)
+    loss_fn_cls: type[nn.Module] = fixed_kwargs.get("loss_fn_cls", nn.MSELoss)
+    epochs: int = fixed_kwargs.get("epochs", 100)
+    batch_size: int = fixed_kwargs.get("batch_size", 16)
+    layer_dims: list[int] | None = fixed_kwargs.get("layer_dims")
+    weight_init: str = fixed_kwargs.get("weight_init", "default")
+    scale: float = fixed_kwargs.get("scale", 1.0)
+    bias_init: str = fixed_kwargs.get("bias_init", "default")
+    dropout_prob: float | None = fixed_kwargs.get("dropout_prob")
+    lr: float = fixed_kwargs.get("lr", 1e-2)
+    params_size: int = fixed_kwargs.get("params_size", 1)
+    random_seed: int | None = fixed_kwargs.get("random_seed")
+    device: DeviceLike | None = fixed_kwargs.get("device")
[…]
