### 2025-11-13 eee1a89
ghrs: report 11-13-2304-C1A2 for alan-turing-institute/autoemulate
