### 2025-11-14 012776d
Fix subclassing and defaults
