### 2025-11-14 df81353
ghrs: report 11-14-2304-5DE8 for alan-turing-institute/autoemulate
