### 2025-11-15 3fda6df
ghrs: report 11-15-2304-E7DF for alan-turing-institute/autoemulate
