### 2025-11-16 592a86d
ghrs: report 11-16-2304-284E for alan-turing-institute/autoemulate
