### 2025-11-17 93d0b50
add support for MSLL per output if have independent predictions
