### 2025-11-17 7e786b5
Ensure bounds ordering is valid
