### 2025-11-17 4dd91d6
ghrs: report 11-17-2304-349C for alan-turing-institute/autoemulate
