### 2025-11-18 e985c8c
Update autoemulate/core/model_selection.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
