### 2025-11-18 ea1130b
Update autoemulate/core/metrics.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
