### 2025-11-18 ce24df4
Update autoemulate/core/metrics.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
