+++ tests/emulators/test_gaussian_process_exact.py
@@ -169,9 +169,9 @@ def test_gp_corr_deterministic_with_seed(sample_data_y1d, new_data_y1d, device):
     new_seed = 43
     set_random_seed(seed)
     model1 = GaussianProcessCorrelated(x, y, device=device)
-    set_random_seed(seed)
-    model2 = GaussianProcessCorrelated(x, y, device=device)
     set_random_seed(new_seed)
+    model2 = GaussianProcessCorrelated(x, y, device=device)
+    set_random_seed(seed)
     model3 = GaussianProcessCorrelated(x, y, device=device)
     model1.fit(x, y)
     pred1 = model1.predict(x2)
