### 2025-11-18 cb35677
fix metric output reshaping
don't use batch dim
