### 2025-11-19 002357d
ghrs: report 11-19-2304-0996 for alan-turing-institute/autoemulate
