### 2025-11-20 24e6f4e
ghrs: report 11-20-2304-412C for alan-turing-institute/autoemulate
