### 2025-11-21 138c5f3
Added Plot Loss and Test
