+++ tests/core/test_plotting.py
@@ -64,3 +64,37 @@ def test_plot_xy():
 def test_calculate_subplot_layout(n_plots, n_cols, expected):
     result = plotting.calculate_subplot_layout(n_plots, n_cols)
     assert result == expected
+
+
+from autoemulate.emulators.polynomials import PolynomialRegression
+from autoemulate.emulators.random_forest import RandomForest
+
+
+@pytest.mark.parametrize(
+    "model_class,should_raise",
+    [
+        (PolynomialRegression, False),
+        (RandomForest, True),
+    ],
+)
+def test_plot_loss(model_class, should_raise):
+    np.random.seed(42)
+    x = np.random.rand(20, 2)
+    y = (x[:, 0] + 2 * x[:, 1] > 1).astype(int)
+
+    model = model_class(x, y)
+    model.fit(x, y)
+
+    if should_raise:
+        with pytest.raises(AttributeError):
+            plotting.plot_loss(model=model, title="Train Loss")
+    else:
+        fig, ax = plotting.plot_loss(model=model, title="Train Loss")
+        assert ax.get_title() == "Train Loss"
+        assert ax.get_xlabel() == "Epochs"
+        assert ax.get_ylabel() == "Train Loss"
+        assert len(model.loss_history) > 0
+        assert np.allclose(
+            ax.get_lines()[0].get_data(),
+            (range(1, len(model.loss_history) + 1), model.loss_history),
+        )
