### 2025-11-23 da57267
ghrs: report 11-23-2304-7625 for alan-turing-institute/autoemulate
