+++ tests/core/test_plotting.py
@@ -71,13 +71,15 @@ from autoemulate.emulators.random_forest import RandomForest
 
 
 @pytest.mark.parametrize(
-    "model_class,should_raise",
+    "model_class,should_raise,title",
     [
-        (PolynomialRegression, False),
-        (RandomForest, True),
+        (PolynomialRegression, False, "Training Curve"),  
+        (RandomForest, True, "My Loss Plot"),             
+        (PolynomialRegression, False, None),             
     ],
 )
-def test_plot_loss(model_class, should_raise):
+
+def test_plot_loss(model_class, should_raise, title):
     np.random.seed(42)
     x = np.random.rand(20, 2)
     y = (x[:, 0] + 2 * x[:, 1] > 1).astype(int)
@@ -87,14 +89,19 @@ def test_plot_loss(model_class, should_raise):
 
     if should_raise:
         with pytest.raises(AttributeError):
-            plotting.plot_loss(model=model, title="Train Loss")
-    else:
-        fig, ax = plotting.plot_loss(model=model, title="Train Loss")
-        assert ax.get_title() == "Train Loss"
-        assert ax.get_xlabel() == "Epochs"
-        assert ax.get_ylabel() == "Train Loss"
-        assert len(model.loss_history) > 0
-        assert np.allclose(
-            ax.get_lines()[0].get_data(),
-            (range(1, len(model.loss_history) + 1), model.loss_history),
-        )
+            plotting.plot_loss(model=model, title=title)
+        return
+
+    fig, ax = plotting.plot_loss(model=model, title=title)
+
+    if title is not None:
+        assert ax.get_title() == title
+
+    assert ax.get_xlabel() == "Epochs"
+    assert ax.get_ylabel() == "Train Loss"
+
+    epochs = np.arange(1, len(model.loss_history) + 1)
+    line_x, line_y = ax.get_lines()[0].get_data()
+
+    assert np.allclose(line_x, epochs)
+    assert np.allclose(line_y, model.loss_history)
