### 2026-01-23 #9113 @crvernon
:wave: @espottesmith are you still interested in taking this submission on as topic editor per your previous message?

> Hey [@crvernon](https://github.com/crvernon) if [@adithirgis](https://github.com/adithirgis) can't step in, I can edit this submission.

If so, @sgreenbury will need to make the new JOSS standards updates before moving to full review.
