### 2026-01-23 #9113 @espottesmith
Yep, would be happy to edit (assuming it meets new JOSS standards)
