### 2026-01-27 #9113 @espottesmith
Hey @sgreenbury, let me know if you have any questions about the new JOSS standards that @crvernon mentioned above.
