### 2026-02-06 #9113 @sgreenbury
We've now completed the revisions required for the updated JOSS standards. The submission should now be ready for review. Thanks!
