### 2026-02-09 #9113 @espottesmith
Alright, @sgreenbury, I do judge your submission to be in line with JOSS guidelines. Now we can start looking for reviewers. It'd help if you could provide a list of names/GitHub usernames; see the authors of the above-listed JOSS papers or our [reviewer database](reviewers.joss.theoj.org/reviewers).
