### 2026-02-12 #9113 @sgreenbury
Thanks @espottesmith - I think the following authors from the JOSS papers mentioned above would be particularly suitable:
- https://github.com/marvinschmitt, https://github.com/stefanradev93 (BayesFlow): neural networks, multiple backends, Bayesian inference and model fitting workflows
- https://github.com/Jammy2211 (PyAutoFit): model building/composition, Bayesian inference and model fitting workflows
- https://github.com/alawryaguila (Multi‑view‑AE): modular PyTorch library with support for many autoencoder models

Happy to suggest more if helpful!
