### 2026-02-13 #9113 @espottesmith
@alawryaguila @stefanradev93 @tensorsofthewall @Jammy2211 @pmeier

Do any of y'all have the time/energy/capacity to give your expertise to this review for the _Journal of Open Source Software_?
