### 2026-03-17 #10087 @pmeier
@espottesmith I've finished my review of the paper and the repository. I'll do the actual software and its documentation later this week.
