### 2026-03-24 #10087 @stefanradev93
I am in the process of finalizing my review and going through the tutorials and the code base. I have three main points regarding the paper itself:

- The title v1.0 collides with the most recent release (v2.0)
- The **State of the Field** coverage is extremely sparse. On the one hand, there are related frameworks out there that are not even mentioned, such as `SMT` (https://github.com/SMTorg/smt), `openturns` (https://github.com/openturns) or `pyapprox` (https://github.com/sandialabs/pyapprox/) for emulation or `Ax` (https://github.com/facebook/ax) for active learning. Moreover, simulation-based inference (SBI) frameworks, such as `bayesflow` (disclaimer: I am a core maintainer, https://github.com/bayesflow-org/bayesflow) and `sbi` (https://github.com/sbi-dev/sbi) can also perform emulation with expressive generative networks by training on `(x, y)` pairs, among comprehensive support for "Bayesian calibration" (i.e., parameter estimation / Bayesian inversion). The paper should use this section to clearly state the tradeoffs and differences between `AutoEmulate` and these other libraries instead of stating what has been updated in this version.
- There are no references or reproducible material concerning the 10 collaborations. The case studieS folder (https://github.com/alan-turing-institute/autoemulate/tree/main/case_studies) has one case study focused only on parameter estimation.
