### 2026-04-13 alan-turing-institute/autoemulate#978 @pmeier
test failures and warnings
_Disclaimer: I'm reviewing the `autoemulate` package as part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/10087). This issue is based on version [`v1.1.1.`](https://github.com/alan-turing-institute/autoemulate/tree/v1.1.1)_

---

`uv run pytest` results in 16 failures and 324 warnings:

```
============================================================= warnings summary ==============================================================
.venv/lib/python3.12/site-packages/pyro/ops/stats.py:514
  /home/philip/git/github.com/autoemulate/.venv/lib/python3.12/site-packages/pyro/ops/stats.py:514: SyntaxWarning: invalid escape sequence '\g'
    """

tests/calibration/test_bayesian_calibration.py::test_hmc_single_output[1-1-10-False]
tests/calibration/test_bayesian_calibration.py::test_hmc_single_output[10-1-10-False]
tests/calibration/test_bayesian_calibration.py::test_hmc_single_output[1-2-10-False]
tests/calibration/test_bayesian_calibration.py::test_hmc_single_output[10-2-10-False]
tests/calibration/test_bayesian_calibration.py::test_hmc_multiple_output[1-1-10-False]
tests/calibration/test_bayesian_calibration.py::test_hmc_multiple_output[10-1-10-False]
tests/calibration/test_bayesian_calibration.py::test_hmc_multiple_output[1-2-10-False]
tests/calibration/test_bayesian_calibration.py::test_hmc_multiple_output[10-2-10-False]
  /home/philip/git/github.com/autoemulate/autoemulate/calibration/bayes.py:189: UserWarning: To copy construct from a tensor, it is recommended to use sourceTensor.detach().clone() or sourceTensor.detach().clone().requires_grad_(True), rather than torch.tensor(sourceTensor).
    else torch.tensor(observation_variance + self.model_discrepancy).to(

tests/core/test_compare.py: 3 warnings
tests/emulators/test_transformed.py: 90 warnings
  /home/philip/git/github.com/autoemulate/autoemulate/emulators/ensemble.py:291: NumericalWarning: cov not p.d. - added 1.0e-04 to the diagonal and symmetrized
    make_positive_definite(

tests/core/test_plotting.py::test_display_figure_jupyter
tests/core/test_plotting.py::test_display_figure_terminal
  /home/philip/git/github.com/autoemulate/autoemulate/core/plotting.py:39: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

tests/emulators/test_grads.py: 2 warnings
tests/emulators/test_transformed.py: 44 warnings
  /home/philip/git/github.com/autoemulate/autoemulate/transforms/base.py:419: NumericalWarning: cov not p.d. - added 1.0e-06 to the diagonal and symmetrized
    make_positive_definite(

tests/emulators/test_grads.py: 1 warning
tests/emulators/test_transformed.py: 37 warnings
  /home/philip/git/github.com/autoemulate/autoemulate/transforms/base.py:419: NumericalWarning: cov not p.d. - added 1.0e-05 to the diagonal and symmetrized
    make_positive_definite(

tests/emulators/test_transformed.py: 31 warnings
  /home/philip/git/github.com/autoemulate/autoemulate/transforms/base.py:419: NumericalWarning: cov not p.d. - added 1.0e-04 to the diagonal and symmetrized
    make_positive_definite(

tests/emulators/test_transformed.py::test_transformed_emulator_100_targets[GaussianProcessCorrelated-x_transforms70-y_transforms70-False]
tests/emulators/test_transformed.py::test_transformed_emulator_100_targets[GaussianProcessCorrelated-x_transforms71-y_transforms71-True]
tests/emulators/test_transformed.py::test_transformed_emulator_100_targets[GaussianProcessCorrelated-x_transforms80-y_transforms80-False]
tests/emulators/test_transformed.py::test_transformed_emulator_100_targets[GaussianProcessCorrelated-x_transforms81-y_transforms81-True]
tests/emulators/test_transformed.py::test_transformed_emulator_100_targets[GaussianProcessCorrelated-x_transforms90-y_transforms90-False]
tests/emulators/test_transformed.py::test_transformed_emulator_100_targets[GaussianProcessCorrelated-x_transforms91-y_transforms91-True]
  /home/philip/git/github.com/autoemulate/autoemulate/emulators/gaussian_process/exact.py:289: NumericalWarning: cov not p.d. - clamped min eigval to 1.0e-03 and symmetrized
    make_positive_definite(

tests/emulators/test_transformed.py: 90 warnings
  /home/philip/git/github.com/autoemulate/autoemulate/emulators/ensemble.py:128: NumericalWarning: cov not p.d. - added 1.0e-04 to the diagonal and symmetrized
    make_positive_definite(

tests/emulators/test_transformed.py::test_inverse_gaussian_and_sample_pca
tests/emulators/test_transformed.py::test_inverse_gaussian_and_sample_vae
tests/transforms/test_transforms.py::test_inverse_sample_gaussian_like[loc1-scale1-expected_shape1]
tests/transforms/test_transforms.py::test_inverse_sample_gaussian_like[loc2-scale2-expected_shape2]
  /home/philip/git/github.com/autoemulate/autoemulate/transforms/base.py:494: NumericalWarning: cov not p.d. - added 1.0e-06 to the diagonal and symmetrized
    make_positive_definite(

tests/simulations/test_base_simulator.py::test_sample[sobol]
  /home/philip/git/github.com/autoemulate/autoemulate/simulations/base.py:187: UserWarning: The balance properties of Sobol' points require n to be a power of 2.
    samples = sampler.random(n=n_samples)

tests/transforms/test_transforms.py::test_inverse_sample_gaussian_like[loc0-scale0-expected_shape0]
  /home/philip/git/github.com/autoemulate/autoemulate/transforms/base.py:503: NumericalWarning: cov not p.d. - added 1.0e-06 to the diagonal and symmetrized
    else make_positive_definite(

tests/transforms/test_transforms_utils.py::test_make_positive_definite[1e-06]
  /home/philip/git/github.com/autoemulate/tests/transforms/test_transforms_utils.py:14: NumericalWarning: cov not p.d. - added 1.0e-05 to the diagonal and symmetrized
    covs = make_positive_definite(covs, clamp_eigvals=True)

tests/transforms/test_transforms_utils.py::test_make_positive_definite[1.0]
  /home/philip/git/github.com/autoemulate/tests/transforms/test_transforms_utils.py:14: NumericalWarning: cov not p.d. - clamped min eigval to 1.0e-03 and symmetrized
    covs = make_positive_definite(covs, clamp_eigvals=True)

tests/transforms/test_transforms_utils.py::test_make_positive_definite[1000000.0]
  /home/philip/git/github.com/autoemulate/tests/transforms/test_transforms_utils.py:14: NumericalWarning: cov not p.d. - clamped min eigval to 1.0e-03 and max eigval to 1.0e+03, then symmetrized
    covs = make_positive_definite(covs, clamp_eigvals=True)

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
========================================================== short test summary info ==========================================================
FAILED tests/calibration/test_history_matching.py::test_run[cuda] - RuntimeError: Expected all tensors to be on the same device, but found at least two devices, cuda:0 and cpu!
FAILED tests/emulators/test_gaussian_process_exact.py::test_tune_gp[cuda-GaussianProcess] - RuntimeError: Failed to tune model GaussianProcess after 5 retries.
FAILED tests/emulators/test_gaussian_process_exact.py::test_tune_gp[cuda-GaussianProcessCorrelated] - RuntimeError: Failed to tune model GaussianProcessCorrelated after 5 retries.
FAILED tests/emulators/test_gaussian_process_exact.py::test_tune_gp[cuda-GaussianProcessMatern32] - RuntimeError: Failed to tune model GaussianProcessMatern32 after 5 retries.
FAILED tests/emulators/test_gaussian_process_exact.py::test_tune_gp[cuda-GaussianProcessMatern52] - RuntimeError: Failed to tune model GaussianProcessMatern52 after 5 retries.
FAILED tests/emulators/test_gaussian_process_exact.py::test_tune_gp[cuda-GaussianProcessRBF] - RuntimeError: Failed to tune model GaussianProcessRBF after 5 retries.
FAILED tests/emulators/test_gaussian_process_exact.py::test_device[cuda-GaussianProcess] - RuntimeError: Deterministic behavior was enabled with either `torch.use_deterministic_algorithms(True)` or `at::Context::setDeterministi...
FAILED tests/emulators/test_gaussian_process_exact.py::test_device[cuda-GaussianProcessCorrelated] - RuntimeError: Deterministic behavior was enabled with either `torch.use_deterministic_algorithms(True)` or `at::Context::setDeterministi...
FAILED tests/emulators/test_gaussian_process_exact.py::test_device[cuda-GaussianProcessMatern32] - RuntimeError: Deterministic behavior was enabled with either `torch.use_deterministic_algorithms(True)` or `at::Context::setDeterministi...
FAILED tests/emulators/test_gaussian_process_exact.py::test_device[cuda-GaussianProcessMatern52] - RuntimeError: Deterministic behavior was enabled with either `torch.use_deterministic_algorithms(True)` or `at::Context::setDeterministi...
FAILED tests/emulators/test_gaussian_process_exact.py::test_device[cuda-GaussianProcessRBF] - RuntimeError: Deterministic behavior was enabled with either `torch.use_deterministic_algorithms(True)` or `at::Context::setDeterministi...
FAILED tests/emulators/test_gaussian_process_exact.py::test_gp_deterministic_with_seed[cuda] - RuntimeError: Deterministic behavior was enabled with either `torch.use_deterministic_algorithms(True)` or `at::Context::setDeterministi...
FAILED tests/emulators/test_gaussian_process_exact.py::test_gp_corr_deterministic_with_seed[cuda] - RuntimeError: Deterministic behavior was enabled with either `torch.use_deterministic_algorithms(True)` or `at::Context::setDeterministi...
FAILED tests/emulators/test_mlp.py::test_tune_mlp[cuda] - RuntimeError: Failed to tune model MLP after 5 retries.
FAILED tests/emulators/test_polynomials.py::test_tune_pr[cuda] - RuntimeError: Failed to tune model PolynomialRegression after 5 retries.
FAILED tests/emulators/test_radial_basis_functions.py::test_tune_rbf[cuda] - RuntimeError: Failed to tune model RadialBasisFunctions after 5 retries.
============================ 16 failed, 1754 passed, 82 skipped, 106 xfailed, 324 warnings in 2097.11s (0:34:57) ============================
```

[Full output](https://github.com/user-attachments/files/26688078/test.log)
