### 2026-04-13 alan-turing-institute/autoemulate#979 @pmeier
benchmark script fails
_Disclaimer: I'm reviewing the `autoemulate` package as part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/10087). This issue is based on version [`v1.1.1.`](https://github.com/alan-turing-institute/autoemulate/tree/v1.1.1)_

---

I've installed `pueue`

https://github.com/alan-turing-institute/autoemulate/blob/24f32d7fa0cc9d167fab3cb1ed1570b85ca7f799/benchmarks/README.md?plain=1#L8

```shell
pueue --version
pueue 4.0.4
```

Running the `run_benchmark.sh` from inside the `benchmarks` folder results in a venv error

```
./run_benchmark.sh
./run_benchmark.sh: line 3: .venv/bin/activate: No such file or directory
```

Running from the project root instead results in what is supposedly a `pueue` error

```
./benchmarks/run_benchmark.sh
Running benchmark for simulator: Epidemic, n_splits: 5, n_iter: 10 100
Error:
   0: Couldn't find a configuration file. Did you start the daemon yet?

Location:
   pueue/src/bin/pueue.rs:61

Backtrace omitted. Run with RUST_BACKTRACE=1 environment variable to display it.
Run with RUST_BACKTRACE=full to include source snippets.
```
