### 2026-04-13 alan-turing-institute/autoemulate#980 @pmeier
logging should not be configured
_Disclaimer: I'm reviewing the `autoemulate` package as part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/10087). This issue is based on version [`v1.1.1.`](https://github.com/alan-turing-institute/autoemulate/tree/v1.1.1)_

---

Quoting the [official Python documentation](https://docs.python.org/3/howto/logging.html#library-config):

> It is strongly advised that you do not add any handlers other than [NullHandler](https://docs.python.org/3/library/logging.handlers.html#logging.NullHandler) to your library’s loggers. This is because the configuration of handlers is the prerogative of the application developer who uses your library. The application developer knows their target audience and what handlers are most appropriate for their application: if you add handlers ‘under the hood’, you might well interfere with their ability to carry out unit tests and deliver logs which suit their requirements.

Yet

https://github.com/alan-turing-institute/autoemulate/blob/24f32d7fa0cc9d167fab3cb1ed1570b85ca7f799/autoemulate/core/logging_config.py#L7-L21
