### 2026-04-13 #10087 @pmeier
@espottesmith I'm sorry for the delay. I did my review today. Apart from the concern about the state of the field section from @stefanradev93 in https://github.com/openjournals/joss-reviews/issues/10087#issuecomment-4118854650 that I want back as well, I also opened three issues on the repository:

- alan-turing-institute/autoemulate#978: automated tests are failing for me and the ones that don't fail emit a ton of warnings. Some of which point to code quality issues that likely can easily be fixed. The tutorials also usually start with `warnings.filterwarnings("ignore")`, which might be harmless, but could also hide more issues.
- alan-turing-institute/autoemulate#979: the repository features a benchmark script that fails to run.
- alan-turing-institute/autoemulate#980: `autoemulate` sees itself as library to the best of my understanding. However, it is configuring logging as if it were an application against official Python suggestion. This is not a show-stopper for me in the context of this review, but I think it is useful to be mentioned.
