### 2026-04-14 #10087 @sgreenbury
Thanks @espottesmith for following up, and thanks to @stefanradev93 and @pmeier for the reviews - we really appreciate the thoughtful feedback.

We've begun updating the manuscript in response and will address the opened issues. We'll post an update here once these are completed.
