### 2026-04-14 alan-turing-institute/autoemulate#979 @radka-j
@all-contributors please add @pmeier for bug
