### 2026-04-29 #10087 @pmeier
Just to get ahead of potential issues: The version under review is v1.1.1 as indicated in the [top comment](https://github.com/openjournals/joss-reviews/issues/10087#issue-3968064033). The fix that already has been merged, went into the main branch for which the latest release is [v2.0.0](https://github.com/alan-turing-institute/autoemulate/releases/tag/v2.0.0) at the time of writing. Meaning, if the patches get released as is, the smallest version could be v2.0.1, which is a major version jump and a jump of three releases in total compared to what was reviewed.

@espottesmith please advise how the next round of review should be handled.
