### 2026-04-30 #10087 @espottesmith
Appreciate this comment @pmeier.

@sgreenbury, it's true that this submission is ostensibly for version 1 of your code (including in the manuscript title), but if your modifications in response to peer review are going into a later version, it would seem to me that that's what reviewers should be reviewing and what should (hopefully, eventually) be published in JOSS.

What are your thoughts?
