### 2026-04-30 #10087 @sgreenbury
Thanks @pmeier and @espottesmith for flagging this.

We bumped the version to v2.0 during the pre-review process (#9113) to clarify stability following the discussion in the thread and later reflected in the pre-review decision: https://github.com/openjournals/joss-reviews/issues/9113#issuecomment-3790586971

We've updated the paper accordingly and are happy for v2.0 to be treated as the version under review with only minor/patch releases applied going forward.
