### 2026-05-01 #10087 @espottesmith
Cool! Makes sense to me. For now, I'll set v2.0.0 as the review version.
