### 2026-08-18 #10087 @crvernon
@sgreenbury - please address the following before I accept this submission for publication:

- [ ] LINE 67: “AutoEmulate design” should be written as “AutoEmulate’s design"
- [ ] LINE 98: The second code example appears to be missing a comma after `x_transforms_list=[[StandardizeTransform]]`
- [ ] LINE 107: “dummy example” is informal for a journal article. Consider “minimal example” or “illustrative example”
- [ ] LINE 122: Expand “OFEME2025” on first use or provide enough identifying information for readers unfamiliar with the event
- [ ] LINE 146: “gaussian process” should be “Gaussian process" which you can do in your bib file by placing characters you wish to maintain formatting for in curly brackets.
- [ ] LINE 152: “bayesian inference” should be “Bayesian inference"
- [ ]  LINE 161: “python” should be “Python"
 - [ ] LINE 179: “gaussian processes” should be “Gaussian processes"
 - [ ] LINE 182: “python package” should be “Python package"

Let me know when these changes have been addressed.
